Remove duplicate definitions, delete unnecessary manual files,
