+++ src/System.ComponentModel.TypeConverter/ref/System.ComponentModel.TypeConverter.csproj
@@ -4,7 +4,6 @@
     <Configurations>netcoreapp-Debug;netcoreapp-Release;uap-Debug;uap-Release</Configurations>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="System.ComponentModel.cs" />
     <Compile Include="System.ComponentModel.manual.cs" />
     <Compile Include="System.ComponentModel.TypeConverter.cs" />
     <Compile Include="System.ComponentModel.TypeConverter.Forwards.cs" />
