--- src/System.ComponentModel.TypeConverter/ref/System.ComponentModel.cs
@@ -1,1492 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.ComponentModel
-{
-    [System.AttributeUsageAttribute((System.AttributeTargets)(128))]
-    public sealed partial class SettingsBindableAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.SettingsBindableAttribute No;
-        public static readonly System.ComponentModel.SettingsBindableAttribute Yes;
-        public SettingsBindableAttribute(bool bindable) { }
-        public bool Bindable { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    [System.ComponentModel.DesignerCategoryAttribute("Component")]
-    [System.ComponentModel.TypeConverter(typeof(ComponentConverter))]
-    public partial class MarshalByValueComponent : System.ComponentModel.IComponent, System.IDisposable, System.IServiceProvider
-    {
-        public MarshalByValueComponent() { }
-        [System.ComponentModel.BrowsableAttribute(false)]
-        [System.ComponentModel.DesignerSerializationVisibilityAttribute((System.ComponentModel.DesignerSerializationVisibility)(0))]
-        public virtual System.ComponentModel.IContainer Container { get { throw null; } }
-        [System.ComponentModel.BrowsableAttribute(false)]
-        [System.ComponentModel.DesignerSerializationVisibilityAttribute((System.ComponentModel.DesignerSerializationVisibility)(0))]
-        public virtual bool DesignMode { get { throw null; } }
-        protected System.ComponentModel.EventHandlerList Events { get { throw null; } }
-        [System.ComponentModel.BrowsableAttribute(false)]
-        [System.ComponentModel.DesignerSerializationVisibilityAttribute((System.ComponentModel.DesignerSerializationVisibility)(0))]
-        public virtual System.ComponentModel.ISite Site { get { throw null; } set { } }
-        public event System.EventHandler Disposed { add { } remove { } }
-        public void Dispose() { }
-        protected virtual void Dispose(bool disposing) { }
-        ~MarshalByValueComponent() { }
-        public virtual object GetService(System.Type service) { throw null; }
-        public override string ToString() { throw null; }
-    }
-    public partial class AddingNewEventArgs : System.EventArgs
-    {
-        public AddingNewEventArgs() { }
-        public AddingNewEventArgs(object newObject) { }
-        public object NewObject { get { throw null; } set { } }
-    }
-    public delegate void AddingNewEventHandler(object sender, System.ComponentModel.AddingNewEventArgs e);
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public sealed partial class AmbientValueAttribute : System.Attribute
-    {
-        public AmbientValueAttribute(bool value) { }
-        public AmbientValueAttribute(byte value) { }
-        public AmbientValueAttribute(char value) { }
-        public AmbientValueAttribute(double value) { }
-        public AmbientValueAttribute(short value) { }
-        public AmbientValueAttribute(int value) { }
-        public AmbientValueAttribute(long value) { }
-        public AmbientValueAttribute(object value) { }
-        public AmbientValueAttribute(float value) { }
-        public AmbientValueAttribute(string value) { }
-        public AmbientValueAttribute(System.Type type, string value) { }
-        public object Value { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public sealed partial class BindableAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.BindableAttribute Default;
-        public static readonly System.ComponentModel.BindableAttribute No;
-        public static readonly System.ComponentModel.BindableAttribute Yes;
-        public BindableAttribute(bool bindable) { }
-        public BindableAttribute(bool bindable, System.ComponentModel.BindingDirection direction) { }
-        public BindableAttribute(System.ComponentModel.BindableSupport flags) { }
-        public BindableAttribute(System.ComponentModel.BindableSupport flags, System.ComponentModel.BindingDirection direction) { }
-        public bool Bindable { get { throw null; } }
-        public System.ComponentModel.BindingDirection Direction { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    public enum BindableSupport
-    {
-        Default = 2,
-        No = 0,
-        Yes = 1,
-    }
-    public enum BindingDirection
-    {
-        OneWay = 0,
-        TwoWay = 1,
-    }
-    public partial class BindingList<T> : System.Collections.ObjectModel.Collection<T>, System.Collections.ICollection, System.Collections.IEnumerable, System.Collections.IList, System.ComponentModel.IBindingList, System.ComponentModel.ICancelAddNew, System.ComponentModel.IRaiseItemChangedEvents
-    {
-        public BindingList() { }
-        public BindingList(System.Collections.Generic.IList<T> list) { }
-        public bool AllowEdit { get { throw null; } set { } }
-        public bool AllowNew { get { throw null; } set { } }
-        public bool AllowRemove { get { throw null; } set { } }
-        protected virtual bool IsSortedCore { get { throw null; } }
-        public bool RaiseListChangedEvents { get { throw null; } set { } }
-        protected virtual System.ComponentModel.ListSortDirection SortDirectionCore { get { throw null; } }
-        protected virtual System.ComponentModel.PropertyDescriptor SortPropertyCore { get { throw null; } }
-        protected virtual bool SupportsChangeNotificationCore { get { throw null; } }
-        protected virtual bool SupportsSearchingCore { get { throw null; } }
-        protected virtual bool SupportsSortingCore { get { throw null; } }
-        bool System.ComponentModel.IBindingList.AllowEdit { get { throw null; } }
-        bool System.ComponentModel.IBindingList.AllowNew { get { throw null; } }
-        bool System.ComponentModel.IBindingList.AllowRemove { get { throw null; } }
-        bool System.ComponentModel.IBindingList.IsSorted { get { throw null; } }
-        System.ComponentModel.ListSortDirection System.ComponentModel.IBindingList.SortDirection { get { throw null; } }
-        System.ComponentModel.PropertyDescriptor System.ComponentModel.IBindingList.SortProperty { get { throw null; } }
-        bool System.ComponentModel.IBindingList.SupportsChangeNotification { get { throw null; } }
-        bool System.ComponentModel.IBindingList.SupportsSearching { get { throw null; } }
-        bool System.ComponentModel.IBindingList.SupportsSorting { get { throw null; } }
-        bool System.ComponentModel.IRaiseItemChangedEvents.RaisesItemChangedEvents { get { throw null; } }
-        public event System.ComponentModel.AddingNewEventHandler AddingNew { add { } remove { } }
-        public event System.ComponentModel.ListChangedEventHandler ListChanged { add { } remove { } }
-        public T AddNew() { throw null; }
-        protected virtual object AddNewCore() { throw null; }
-        protected virtual void ApplySortCore(System.ComponentModel.PropertyDescriptor prop, System.ComponentModel.ListSortDirection direction) { }
-        public virtual void CancelNew(int itemIndex) { }
-        protected override void ClearItems() { }
-        public virtual void EndNew(int itemIndex) { }
-        protected virtual int FindCore(System.ComponentModel.PropertyDescriptor prop, object key) { throw null; }
-        protected override void InsertItem(int index, T item) { }
-        protected virtual void OnAddingNew(System.ComponentModel.AddingNewEventArgs e) { }
-        protected virtual void OnListChanged(System.ComponentModel.ListChangedEventArgs e) { }
-        protected override void RemoveItem(int index) { }
-        protected virtual void RemoveSortCore() { }
-        public void ResetBindings() { }
-        public void ResetItem(int position) { }
-        protected override void SetItem(int index, T item) { }
-        void System.ComponentModel.IBindingList.AddIndex(System.ComponentModel.PropertyDescriptor prop) { }
-        object System.ComponentModel.IBindingList.AddNew() { throw null; }
-        void System.ComponentModel.IBindingList.ApplySort(System.ComponentModel.PropertyDescriptor prop, System.ComponentModel.ListSortDirection direction) { }
-        int System.ComponentModel.IBindingList.Find(System.ComponentModel.PropertyDescriptor prop, object key) { throw null; }
-        void System.ComponentModel.IBindingList.RemoveIndex(System.ComponentModel.PropertyDescriptor prop) { }
-        void System.ComponentModel.IBindingList.RemoveSort() { }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public sealed partial class ComplexBindingPropertiesAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.ComplexBindingPropertiesAttribute Default;
-        public ComplexBindingPropertiesAttribute() { }
-        public ComplexBindingPropertiesAttribute(string dataSource) { }
-        public ComplexBindingPropertiesAttribute(string dataSource, string dataMember) { }
-        public string DataMember { get { throw null; } }
-        public string DataSource { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    public partial class ComponentConverter : System.ComponentModel.ReferenceConverter
-    {
-        public ComponentConverter(System.Type type) : base(default(System.Type)) { }
-        public override System.ComponentModel.PropertyDescriptorCollection GetProperties(System.ComponentModel.ITypeDescriptorContext context, object value, System.Attribute[] attributes) { throw null; }
-        public override bool GetPropertiesSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-    }
-    public abstract partial class ComponentEditor
-    {
-        protected ComponentEditor() { }
-        public abstract bool EditComponent(System.ComponentModel.ITypeDescriptorContext context, object component);
-        public bool EditComponent(object component) { throw null; }
-    }
-    public partial class ComponentResourceManager : System.Resources.ResourceManager
-    {
-        public ComponentResourceManager() { }
-        public ComponentResourceManager(System.Type t) : base(t) { }
-        public void ApplyResources(object value, string objectName) { }
-        public virtual void ApplyResources(object value, string objectName, System.Globalization.CultureInfo culture) { }
-    }
-    public partial class Container : System.ComponentModel.IContainer, System.IDisposable
-    {
-        public Container() { }
-        public virtual System.ComponentModel.ComponentCollection Components { get { throw null; } }
-        public virtual void Add(System.ComponentModel.IComponent component) { }
-        public virtual void Add(System.ComponentModel.IComponent component, string name) { }
-        protected virtual System.ComponentModel.ISite CreateSite(System.ComponentModel.IComponent component, string name) { throw null; }
-        public void Dispose() { }
-        protected virtual void Dispose(bool disposing) { }
-        ~Container() { }
-        protected virtual object GetService(System.Type service) { throw null; }
-        public virtual void Remove(System.ComponentModel.IComponent component) { }
-        protected void RemoveWithoutUnsiting(System.ComponentModel.IComponent component) { }
-        protected virtual void ValidateName(System.ComponentModel.IComponent component, string name) { }
-    }
-    public abstract partial class ContainerFilterService
-    {
-        protected ContainerFilterService() { }
-        public virtual System.ComponentModel.ComponentCollection FilterComponents(System.ComponentModel.ComponentCollection components) { throw null; }
-    }
-    public partial class CultureInfoConverter : System.ComponentModel.TypeConverter
-    {
-        public CultureInfoConverter() { }
-        public override bool CanConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Type sourceType) { throw null; }
-        public override bool CanConvertTo(System.ComponentModel.ITypeDescriptorContext context, System.Type destinationType) { throw null; }
-        public override object ConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value) { throw null; }
-        public override object ConvertTo(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value, System.Type destinationType) { throw null; }
-        protected virtual string GetCultureName(System.Globalization.CultureInfo culture) { throw null; }
-        public override System.ComponentModel.TypeConverter.StandardValuesCollection GetStandardValues(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-        public override bool GetStandardValuesExclusive(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-        public override bool GetStandardValuesSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public sealed partial class DataObjectAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.DataObjectAttribute DataObject;
-        public static readonly System.ComponentModel.DataObjectAttribute Default;
-        public static readonly System.ComponentModel.DataObjectAttribute NonDataObject;
-        public DataObjectAttribute() { }
-        public DataObjectAttribute(bool isDataObject) { }
-        public bool IsDataObject { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(128))]
-    public sealed partial class DataObjectFieldAttribute : System.Attribute
-    {
-        public DataObjectFieldAttribute(bool primaryKey) { }
-        public DataObjectFieldAttribute(bool primaryKey, bool isIdentity) { }
-        public DataObjectFieldAttribute(bool primaryKey, bool isIdentity, bool isNullable) { }
-        public DataObjectFieldAttribute(bool primaryKey, bool isIdentity, bool isNullable, int length) { }
-        public bool IsIdentity { get { throw null; } }
-        public bool IsNullable { get { throw null; } }
-        public int Length { get { throw null; } }
-        public bool PrimaryKey { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(64))]
-    public sealed partial class DataObjectMethodAttribute : System.Attribute
-    {
-        public DataObjectMethodAttribute(System.ComponentModel.DataObjectMethodType methodType) { }
-        public DataObjectMethodAttribute(System.ComponentModel.DataObjectMethodType methodType, bool isDefault) { }
-        public bool IsDefault { get { throw null; } }
-        public System.ComponentModel.DataObjectMethodType MethodType { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool Match(object obj) { throw null; }
-    }
-    public enum DataObjectMethodType
-    {
-        Delete = 4,
-        Fill = 0,
-        Insert = 3,
-        Select = 1,
-        Update = 2,
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public sealed partial class DefaultBindingPropertyAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.DefaultBindingPropertyAttribute Default;
-        public DefaultBindingPropertyAttribute() { }
-        public DefaultBindingPropertyAttribute(string name) { }
-        public string Name { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(1028), AllowMultiple = true, Inherited = true)]
-    public sealed partial class DesignerAttribute : System.Attribute
-    {
-        public DesignerAttribute(string designerTypeName) { }
-        public DesignerAttribute(string designerTypeName, string designerBaseTypeName) { }
-        public DesignerAttribute(string designerTypeName, System.Type designerBaseType) { }
-        public DesignerAttribute(System.Type designerType) { }
-        public DesignerAttribute(System.Type designerType, System.Type designerBaseType) { }
-        public string DesignerBaseTypeName { get { throw null; } }
-        public string DesignerTypeName { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(1028))]
-    public sealed partial class DesignTimeVisibleAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.DesignTimeVisibleAttribute Default;
-        public static readonly System.ComponentModel.DesignTimeVisibleAttribute No;
-        public static readonly System.ComponentModel.DesignTimeVisibleAttribute Yes;
-        public DesignTimeVisibleAttribute() { }
-        public DesignTimeVisibleAttribute(bool visible) { }
-        public bool Visible { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767), AllowMultiple = true, Inherited = true)]
-    public sealed partial class EditorAttribute : System.Attribute
-    {
-        public EditorAttribute() { }
-        public EditorAttribute(string typeName, string baseTypeName) { }
-        public EditorAttribute(string typeName, System.Type baseType) { }
-        public EditorAttribute(System.Type type, System.Type baseType) { }
-        public string EditorBaseTypeName { get { throw null; } }
-        public string EditorTypeName { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    public partial class ExpandableObjectConverter : System.ComponentModel.TypeConverter
-    {
-        public ExpandableObjectConverter() { }
-        public override System.ComponentModel.PropertyDescriptorCollection GetProperties(System.ComponentModel.ITypeDescriptorContext context, object value, System.Attribute[] attributes) { throw null; }
-        public override bool GetPropertiesSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-    }
-    public partial interface IBindingList : System.Collections.ICollection, System.Collections.IEnumerable, System.Collections.IList
-    {
-        bool AllowEdit { get; }
-        bool AllowNew { get; }
-        bool AllowRemove { get; }
-        bool IsSorted { get; }
-        System.ComponentModel.ListSortDirection SortDirection { get; }
-        System.ComponentModel.PropertyDescriptor SortProperty { get; }
-        bool SupportsChangeNotification { get; }
-        bool SupportsSearching { get; }
-        bool SupportsSorting { get; }
-        event System.ComponentModel.ListChangedEventHandler ListChanged;
-        void AddIndex(System.ComponentModel.PropertyDescriptor property);
-        object AddNew();
-        void ApplySort(System.ComponentModel.PropertyDescriptor property, System.ComponentModel.ListSortDirection direction);
-        int Find(System.ComponentModel.PropertyDescriptor property, object key);
-        void RemoveIndex(System.ComponentModel.PropertyDescriptor property);
-        void RemoveSort();
-    }
-    public partial interface IBindingListView : System.Collections.ICollection, System.Collections.IEnumerable, System.Collections.IList, System.ComponentModel.IBindingList
-    {
-        string Filter { get; set; }
-        System.ComponentModel.ListSortDescriptionCollection SortDescriptions { get; }
-        bool SupportsAdvancedSorting { get; }
-        bool SupportsFiltering { get; }
-        void ApplySort(System.ComponentModel.ListSortDescriptionCollection sorts);
-        void RemoveFilter();
-    }
-    public partial interface ICancelAddNew
-    {
-        void CancelNew(int itemIndex);
-        void EndNew(int itemIndex);
-    }
-    [System.ObsoleteAttribute("This interface has been deprecated. Add a TypeDescriptionProvider to handle type TypeDescriptor.ComObjectType instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-    public partial interface IComNativeDescriptorHandler
-    {
-        System.ComponentModel.AttributeCollection GetAttributes(object component);
-        string GetClassName(object component);
-        System.ComponentModel.TypeConverter GetConverter(object component);
-        System.ComponentModel.EventDescriptor GetDefaultEvent(object component);
-        System.ComponentModel.PropertyDescriptor GetDefaultProperty(object component);
-        object GetEditor(object component, System.Type baseEditorType);
-        System.ComponentModel.EventDescriptorCollection GetEvents(object component);
-        System.ComponentModel.EventDescriptorCollection GetEvents(object component, System.Attribute[] attributes);
-        string GetName(object component);
-        System.ComponentModel.PropertyDescriptorCollection GetProperties(object component, System.Attribute[] attributes);
-        object GetPropertyValue(object component, int dispid, ref bool success);
-        object GetPropertyValue(object component, string propertyName, ref bool success);
-    }
-    public partial interface IDataErrorInfo
-    {
-        string Error { get; }
-        string this[string columnName] { get; }
-    }
-    public partial interface IIntellisenseBuilder
-    {
-        string Name { get; }
-        bool Show(string language, string value, ref string newValue);
-    }
-    public partial interface INestedContainer : System.ComponentModel.IContainer, System.IDisposable
-    {
-        System.ComponentModel.IComponent Owner { get; }
-    }
-    public partial interface INestedSite : System.ComponentModel.ISite, System.IServiceProvider
-    {
-        string FullName { get; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(896))]
-    public sealed partial class InheritanceAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.InheritanceAttribute Default;
-        public static readonly System.ComponentModel.InheritanceAttribute Inherited;
-        public static readonly System.ComponentModel.InheritanceAttribute InheritedReadOnly;
-        public static readonly System.ComponentModel.InheritanceAttribute NotInherited;
-        public InheritanceAttribute() { }
-        public InheritanceAttribute(System.ComponentModel.InheritanceLevel inheritanceLevel) { }
-        public System.ComponentModel.InheritanceLevel InheritanceLevel { get { throw null; } }
-        public override bool Equals(object value) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-        public override string ToString() { throw null; }
-    }
-    public enum InheritanceLevel
-    {
-        Inherited = 1,
-        InheritedReadOnly = 2,
-        NotInherited = 3,
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public partial class InstallerTypeAttribute : System.Attribute
-    {
-        public InstallerTypeAttribute(string typeName) { }
-        public InstallerTypeAttribute(System.Type installerType) { }
-        public virtual System.Type InstallerType { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    public abstract partial class InstanceCreationEditor
-    {
-        protected InstanceCreationEditor() { }
-        public virtual string Text { get { throw null; } }
-        public abstract object CreateInstance(System.ComponentModel.ITypeDescriptorContext context, System.Type instanceType);
-    }
-    public partial interface IRaiseItemChangedEvents
-    {
-        bool RaisesItemChangedEvents { get; }
-    }
-    public partial interface ISupportInitializeNotification : System.ComponentModel.ISupportInitialize
-    {
-        bool IsInitialized { get; }
-        event System.EventHandler Initialized;
-    }
-    public abstract partial class License : System.IDisposable
-    {
-        protected License() { }
-        public abstract string LicenseKey { get; }
-        public abstract void Dispose();
-    }
-    public partial class LicenseContext : System.IServiceProvider
-    {
-        public LicenseContext() { }
-        public virtual System.ComponentModel.LicenseUsageMode UsageMode { get { throw null; } }
-        public virtual string GetSavedLicenseKey(System.Type type, System.Reflection.Assembly resourceAssembly) { throw null; }
-        public virtual object GetService(System.Type type) { throw null; }
-        public virtual void SetSavedLicenseKey(System.Type type, string key) { }
-    }
-    public partial class LicenseException : System.SystemException
-    {
-        protected LicenseException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
-        public LicenseException(System.Type type) { }
-        public LicenseException(System.Type type, object instance) { }
-        public LicenseException(System.Type type, object instance, string message) { }
-        public LicenseException(System.Type type, object instance, string message, System.Exception innerException) { }
-        public System.Type LicensedType { get { throw null; } }
-        public override void GetObjectData(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
-    }
-    public sealed partial class LicenseManager
-    {
-        internal LicenseManager() { }
-        public static System.ComponentModel.LicenseContext CurrentContext { get { throw null; } set { } }
-        public static System.ComponentModel.LicenseUsageMode UsageMode { get { throw null; } }
-        public static object CreateWithContext(System.Type type, System.ComponentModel.LicenseContext creationContext) { throw null; }
-        public static object CreateWithContext(System.Type type, System.ComponentModel.LicenseContext creationContext, object[] args) { throw null; }
-        public static bool IsLicensed(System.Type type) { throw null; }
-        public static bool IsValid(System.Type type) { throw null; }
-        public static bool IsValid(System.Type type, object instance, out System.ComponentModel.License license) { throw null; }
-        public static void LockContext(object contextUser) { }
-        public static void UnlockContext(object contextUser) { }
-        public static void Validate(System.Type type) { }
-        public static System.ComponentModel.License Validate(System.Type type, object instance) { throw null; }
-    }
-    public abstract partial class LicenseProvider
-    {
-        protected LicenseProvider() { }
-        public abstract System.ComponentModel.License GetLicense(System.ComponentModel.LicenseContext context, System.Type type, object instance, bool allowExceptions);
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4), AllowMultiple = false, Inherited = false)]
-    public sealed partial class LicenseProviderAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.LicenseProviderAttribute Default;
-        public LicenseProviderAttribute() { }
-        public LicenseProviderAttribute(string typeName) { }
-        public LicenseProviderAttribute(System.Type type) { }
-        public System.Type LicenseProvider { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-        public override bool Equals(object value) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    public enum LicenseUsageMode
-    {
-        Designtime = 1,
-        Runtime = 0,
-    }
-    public partial class LicFileLicenseProvider : System.ComponentModel.LicenseProvider
-    {
-        public LicFileLicenseProvider() { }
-        protected virtual string GetKey(System.Type type) { throw null; }
-        public override System.ComponentModel.License GetLicense(System.ComponentModel.LicenseContext context, System.Type type, object instance, bool allowExceptions) { throw null; }
-        protected virtual bool IsKeyValid(string key, System.Type type) { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public sealed partial class ListBindableAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.ListBindableAttribute Default;
-        public static readonly System.ComponentModel.ListBindableAttribute No;
-        public static readonly System.ComponentModel.ListBindableAttribute Yes;
-        public ListBindableAttribute(bool listBindable) { }
-        public ListBindableAttribute(System.ComponentModel.BindableSupport flags) { }
-        public bool ListBindable { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    public partial class ListChangedEventArgs : System.EventArgs
-    {
-        public ListChangedEventArgs(System.ComponentModel.ListChangedType listChangedType, System.ComponentModel.PropertyDescriptor propDesc) { }
-        public ListChangedEventArgs(System.ComponentModel.ListChangedType listChangedType, int newIndex) { }
-        public ListChangedEventArgs(System.ComponentModel.ListChangedType listChangedType, int newIndex, System.ComponentModel.PropertyDescriptor propDesc) { }
-        public ListChangedEventArgs(System.ComponentModel.ListChangedType listChangedType, int newIndex, int oldIndex) { }
-        public System.ComponentModel.ListChangedType ListChangedType { get { throw null; } }
-        public int NewIndex { get { throw null; } }
-        public int OldIndex { get { throw null; } }
-        public System.ComponentModel.PropertyDescriptor PropertyDescriptor { get { throw null; } }
-    }
-    public delegate void ListChangedEventHandler(object sender, System.ComponentModel.ListChangedEventArgs e);
-    public enum ListChangedType
-    {
-        ItemAdded = 1,
-        ItemChanged = 4,
-        ItemDeleted = 2,
-        ItemMoved = 3,
-        PropertyDescriptorAdded = 5,
-        PropertyDescriptorChanged = 7,
-        PropertyDescriptorDeleted = 6,
-        Reset = 0,
-    }
-    public partial class ListSortDescription
-    {
-        public ListSortDescription(System.ComponentModel.PropertyDescriptor property, System.ComponentModel.ListSortDirection direction) { }
-        public System.ComponentModel.PropertyDescriptor PropertyDescriptor { get { throw null; } set { } }
-        public System.ComponentModel.ListSortDirection SortDirection { get { throw null; } set { } }
-    }
-    public partial class ListSortDescriptionCollection : System.Collections.ICollection, System.Collections.IEnumerable, System.Collections.IList
-    {
-        public ListSortDescriptionCollection() { }
-        public ListSortDescriptionCollection(System.ComponentModel.ListSortDescription[] sorts) { }
-        public int Count { get { throw null; } }
-        public System.ComponentModel.ListSortDescription this[int index] { get { throw null; } set { } }
-        bool System.Collections.ICollection.IsSynchronized { get { throw null; } }
-        object System.Collections.ICollection.SyncRoot { get { throw null; } }
-        bool System.Collections.IList.IsFixedSize { get { throw null; } }
-        bool System.Collections.IList.IsReadOnly { get { throw null; } }
-        object System.Collections.IList.this[int index] { get { throw null; } set { } }
-        public bool Contains(object value) { throw null; }
-        public void CopyTo(System.Array array, int index) { }
-        public int IndexOf(object value) { throw null; }
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() { throw null; }
-        int System.Collections.IList.Add(object value) { throw null; }
-        void System.Collections.IList.Clear() { }
-        void System.Collections.IList.Insert(int index, object value) { }
-        void System.Collections.IList.Remove(object value) { }
-        void System.Collections.IList.RemoveAt(int index) { }
-    }
-    public enum ListSortDirection
-    {
-        Ascending = 0,
-        Descending = 1,
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public sealed partial class LookupBindingPropertiesAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.LookupBindingPropertiesAttribute Default;
-        public LookupBindingPropertiesAttribute() { }
-        public LookupBindingPropertiesAttribute(string dataSource, string displayMember, string valueMember, string lookupMember) { }
-        public string DataSource { get { throw null; } }
-        public string DisplayMember { get { throw null; } }
-        public string LookupMember { get { throw null; } }
-        public string ValueMember { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-    }
-    public partial class MaskedTextProvider : System.ICloneable
-    {
-        public MaskedTextProvider(string mask) { }
-        public MaskedTextProvider(string mask, bool restrictToAscii) { }
-        public MaskedTextProvider(string mask, char passwordChar, bool allowPromptAsInput) { }
-        public MaskedTextProvider(string mask, System.Globalization.CultureInfo culture) { }
-        public MaskedTextProvider(string mask, System.Globalization.CultureInfo culture, bool restrictToAscii) { }
-        public MaskedTextProvider(string mask, System.Globalization.CultureInfo culture, bool allowPromptAsInput, char promptChar, char passwordChar, bool restrictToAscii) { }
-        public MaskedTextProvider(string mask, System.Globalization.CultureInfo culture, char passwordChar, bool allowPromptAsInput) { }
-        public bool AllowPromptAsInput { get { throw null; } }
-        public bool AsciiOnly { get { throw null; } }
-        public int AssignedEditPositionCount { get { throw null; } }
-        public int AvailableEditPositionCount { get { throw null; } }
-        public System.Globalization.CultureInfo Culture { get { throw null; } }
-        public static char DefaultPasswordChar { get { throw null; } }
-        public int EditPositionCount { get { throw null; } }
-        public System.Collections.IEnumerator EditPositions { get { throw null; } }
-        public bool IncludeLiterals { get { throw null; } set { } }
-        public bool IncludePrompt { get { throw null; } set { } }
-        public static int InvalidIndex { get { throw null; } }
-        public bool IsPassword { get { throw null; } set { } }
-        public char this[int index] { get { throw null; } }
-        public int LastAssignedPosition { get { throw null; } }
-        public int Length { get { throw null; } }
-        public string Mask { get { throw null; } }
-        public bool MaskCompleted { get { throw null; } }
-        public bool MaskFull { get { throw null; } }
-        public char PasswordChar { get { throw null; } set { } }
-        public char PromptChar { get { throw null; } set { } }
-        public bool ResetOnPrompt { get { throw null; } set { } }
-        public bool ResetOnSpace { get { throw null; } set { } }
-        public bool SkipLiterals { get { throw null; } set { } }
-        public bool Add(char input) { throw null; }
-        public bool Add(char input, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Add(string input) { throw null; }
-        public bool Add(string input, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public void Clear() { }
-        public void Clear(out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public object Clone() { throw null; }
-        public int FindAssignedEditPositionFrom(int position, bool direction) { throw null; }
-        public int FindAssignedEditPositionInRange(int startPosition, int endPosition, bool direction) { throw null; }
-        public int FindEditPositionFrom(int position, bool direction) { throw null; }
-        public int FindEditPositionInRange(int startPosition, int endPosition, bool direction) { throw null; }
-        public int FindNonEditPositionFrom(int position, bool direction) { throw null; }
-        public int FindNonEditPositionInRange(int startPosition, int endPosition, bool direction) { throw null; }
-        public int FindUnassignedEditPositionFrom(int position, bool direction) { throw null; }
-        public int FindUnassignedEditPositionInRange(int startPosition, int endPosition, bool direction) { throw null; }
-        public static bool GetOperationResultFromHint(System.ComponentModel.MaskedTextResultHint hint) { throw null; }
-        public bool InsertAt(char input, int position) { throw null; }
-        public bool InsertAt(char input, int position, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool InsertAt(string input, int position) { throw null; }
-        public bool InsertAt(string input, int position, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool IsAvailablePosition(int position) { throw null; }
-        public bool IsEditPosition(int position) { throw null; }
-        public static bool IsValidInputChar(char c) { throw null; }
-        public static bool IsValidMaskChar(char c) { throw null; }
-        public static bool IsValidPasswordChar(char c) { throw null; }
-        public bool Remove() { throw null; }
-        public bool Remove(out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool RemoveAt(int position) { throw null; }
-        public bool RemoveAt(int startPosition, int endPosition) { throw null; }
-        public bool RemoveAt(int startPosition, int endPosition, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Replace(char input, int position) { throw null; }
-        public bool Replace(char input, int startPosition, int endPosition, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Replace(char input, int position, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Replace(string input, int position) { throw null; }
-        public bool Replace(string input, int startPosition, int endPosition, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Replace(string input, int position, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public bool Set(string input) { throw null; }
-        public bool Set(string input, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-        public string ToDisplayString() { throw null; }
-        public override string ToString() { throw null; }
-        public string ToString(bool ignorePasswordChar) { throw null; }
-        public string ToString(bool includePrompt, bool includeLiterals) { throw null; }
-        public string ToString(bool ignorePasswordChar, bool includePrompt, bool includeLiterals, int startPosition, int length) { throw null; }
-        public string ToString(bool includePrompt, bool includeLiterals, int startPosition, int length) { throw null; }
-        public string ToString(bool ignorePasswordChar, int startPosition, int length) { throw null; }
-        public string ToString(int startPosition, int length) { throw null; }
-        public bool VerifyChar(char input, int position, out System.ComponentModel.MaskedTextResultHint hint) { throw null; }
-        public bool VerifyEscapeChar(char input, int position) { throw null; }
-        public bool VerifyString(string input) { throw null; }
-        public bool VerifyString(string input, out int testPosition, out System.ComponentModel.MaskedTextResultHint resultHint) { throw null; }
-    }
-    public enum MaskedTextResultHint
-    {
-        AlphanumericCharacterExpected = -2,
-        AsciiCharacterExpected = -1,
-        CharacterEscaped = 1,
-        DigitExpected = -3,
-        InvalidInput = -51,
-        LetterExpected = -4,
-        NoEffect = 2,
-        NonEditPosition = -54,
-        PositionOutOfRange = -55,
-        PromptCharNotAllowed = -52,
-        SideEffect = 3,
-        SignedDigitExpected = -5,
-        Success = 4,
-        UnavailableEditPosition = -53,
-        Unknown = 0,
-    }
-    public partial class NestedContainer : System.ComponentModel.Container, System.ComponentModel.IContainer, System.ComponentModel.INestedContainer, System.IDisposable
-    {
-        public NestedContainer(System.ComponentModel.IComponent owner) { }
-        public System.ComponentModel.IComponent Owner { get { throw null; } }
-        protected virtual string OwnerName { get { throw null; } }
-        protected override System.ComponentModel.ISite CreateSite(System.ComponentModel.IComponent component, string name) { throw null; }
-        protected override void Dispose(bool disposing) { }
-        protected override object GetService(System.Type service) { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public sealed partial class PasswordPropertyTextAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.PasswordPropertyTextAttribute Default;
-        public static readonly System.ComponentModel.PasswordPropertyTextAttribute No;
-        public static readonly System.ComponentModel.PasswordPropertyTextAttribute Yes;
-        public PasswordPropertyTextAttribute() { }
-        public PasswordPropertyTextAttribute(bool password) { }
-        public bool Password { get { throw null; } }
-        public override bool Equals(object o) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public partial class PropertyTabAttribute : System.Attribute
-    {
-        public PropertyTabAttribute() { }
-        public PropertyTabAttribute(string tabClassName) { }
-        public PropertyTabAttribute(string tabClassName, System.ComponentModel.PropertyTabScope tabScope) { }
-        public PropertyTabAttribute(System.Type tabClass) { }
-        public PropertyTabAttribute(System.Type tabClass, System.ComponentModel.PropertyTabScope tabScope) { }
-        public System.Type[] TabClasses { get { throw null; } }
-        protected string[] TabClassNames { get { throw null; } }
-        public System.ComponentModel.PropertyTabScope[] TabScopes { get { throw null; } }
-        public bool Equals(System.ComponentModel.PropertyTabAttribute other) { throw null; }
-        public override bool Equals(object other) { throw null; }
-        public override int GetHashCode() { throw null; }
-        protected void InitializeArrays(string[] tabClassNames, System.ComponentModel.PropertyTabScope[] tabScopes) { }
-        protected void InitializeArrays(System.Type[] tabClasses, System.ComponentModel.PropertyTabScope[] tabScopes) { }
-    }
-    public enum PropertyTabScope
-    {
-        Component = 3,
-        Document = 2,
-        Global = 1,
-        Static = 0,
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(128))]
-    [System.ObsoleteAttribute("Use System.ComponentModel.SettingsBindableAttribute instead to work with the new settings model.")]
-    public partial class RecommendedAsConfigurableAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.RecommendedAsConfigurableAttribute Default;
-        public static readonly System.ComponentModel.RecommendedAsConfigurableAttribute No;
-        public static readonly System.ComponentModel.RecommendedAsConfigurableAttribute Yes;
-        public RecommendedAsConfigurableAttribute(bool recommendedAsConfigurable) { }
-        public bool RecommendedAsConfigurable { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    public partial class ReferenceConverter : System.ComponentModel.TypeConverter
-    {
-        public ReferenceConverter(System.Type type) { }
-        public override bool CanConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Type sourceType) { throw null; }
-        public override object ConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value) { throw null; }
-        public override object ConvertTo(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value, System.Type destinationType) { throw null; }
-        public override System.ComponentModel.TypeConverter.StandardValuesCollection GetStandardValues(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-        public override bool GetStandardValuesExclusive(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-        public override bool GetStandardValuesSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
-        protected virtual bool IsValueAllowed(System.ComponentModel.ITypeDescriptorContext context, object value) { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4))]
-    public partial class RunInstallerAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.RunInstallerAttribute Default;
-        public static readonly System.ComponentModel.RunInstallerAttribute No;
-        public static readonly System.ComponentModel.RunInstallerAttribute Yes;
-        public RunInstallerAttribute(bool runInstaller) { }
-        public bool RunInstaller { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    public static partial class SyntaxCheck
-    {
-        public static bool CheckMachineName(string value) { throw null; }
-        public static bool CheckPath(string value) { throw null; }
-        public static bool CheckRootedPath(string value) { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767))]
-    public partial class ToolboxItemAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.ToolboxItemAttribute Default;
-        public static readonly System.ComponentModel.ToolboxItemAttribute None;
-        public ToolboxItemAttribute(bool defaultType) { }
-        public ToolboxItemAttribute(string toolboxItemTypeName) { }
-        public ToolboxItemAttribute(System.Type toolboxItemType) { }
-        public System.Type ToolboxItemType { get { throw null; } }
-        public string ToolboxItemTypeName { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4), AllowMultiple = true, Inherited = true)]
-    public sealed partial class ToolboxItemFilterAttribute : System.Attribute
-    {
-        public ToolboxItemFilterAttribute(string filterString) { }
-        public ToolboxItemFilterAttribute(string filterString, System.ComponentModel.ToolboxItemFilterType filterType) { }
-        public string FilterString { get { throw null; } }
-        public System.ComponentModel.ToolboxItemFilterType FilterType { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool Match(object obj) { throw null; }
-        public override string ToString() { throw null; }
-    }
-    public enum ToolboxItemFilterType
-    {
-        Allow = 0,
-        Custom = 1,
-        Prevent = 2,
-        Require = 3,
-    }
-    public partial class WarningException : System.SystemException
-    {
-        public WarningException() { }
-        protected WarningException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
-        public WarningException(string message) { }
-        public WarningException(string message, System.Exception innerException) { }
-        public WarningException(string message, string helpUrl) { }
-        public WarningException(string message, string helpUrl, string helpTopic) { }
-        public string HelpTopic { get { throw null; } }
-        public string HelpUrl { get { throw null; } }
-        public override void GetObjectData(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
-    }
-}
-namespace System.ComponentModel.Design
-{
-    public partial class ActiveDesignerEventArgs : System.EventArgs
-    {
-        public ActiveDesignerEventArgs(System.ComponentModel.Design.IDesignerHost oldDesigner, System.ComponentModel.Design.IDesignerHost newDesigner) { }
-        public System.ComponentModel.Design.IDesignerHost NewDesigner { get { throw null; } }
-        public System.ComponentModel.Design.IDesignerHost OldDesigner { get { throw null; } }
-    }
-    public delegate void ActiveDesignerEventHandler(object sender, System.ComponentModel.Design.ActiveDesignerEventArgs e);
-    public partial class CheckoutException : System.Runtime.InteropServices.ExternalException
-    {
-        public static readonly System.ComponentModel.Design.CheckoutException Canceled;
-        public CheckoutException() { }
-        protected CheckoutException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
-        public CheckoutException(string message) { }
-        public CheckoutException(string message, System.Exception innerException) { }
-        public CheckoutException(string message, int errorCode) { }
-    }
-    public partial class CommandID
-    {
-        public CommandID(System.Guid menuGroup, int commandID) { }
-        public virtual System.Guid Guid { get { throw null; } }
-        public virtual int ID { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override string ToString() { throw null; }
-    }
-    public sealed partial class ComponentChangedEventArgs : System.EventArgs
-    {
-        public ComponentChangedEventArgs(object component, System.ComponentModel.MemberDescriptor member, object oldValue, object newValue) { }
-        public object Component { get { throw null; } }
-        public System.ComponentModel.MemberDescriptor Member { get { throw null; } }
-        public object NewValue { get { throw null; } }
-        public object OldValue { get { throw null; } }
-    }
-    public delegate void ComponentChangedEventHandler(object sender, System.ComponentModel.Design.ComponentChangedEventArgs e);
-    public sealed partial class ComponentChangingEventArgs : System.EventArgs
-    {
-        public ComponentChangingEventArgs(object component, System.ComponentModel.MemberDescriptor member) { }
-        public object Component { get { throw null; } }
-        public System.ComponentModel.MemberDescriptor Member { get { throw null; } }
-    }
-    public delegate void ComponentChangingEventHandler(object sender, System.ComponentModel.Design.ComponentChangingEventArgs e);
-    public partial class ComponentEventArgs : System.EventArgs
-    {
-        public ComponentEventArgs(System.ComponentModel.IComponent component) { }
-        public virtual System.ComponentModel.IComponent Component { get { throw null; } }
-    }
-    public delegate void ComponentEventHandler(object sender, System.ComponentModel.Design.ComponentEventArgs e);
-    public partial class ComponentRenameEventArgs : System.EventArgs
-    {
-        public ComponentRenameEventArgs(object component, string oldName, string newName) { }
-        public object Component { get { throw null; } }
-        public virtual string NewName { get { throw null; } }
-        public virtual string OldName { get { throw null; } }
-    }
-    public delegate void ComponentRenameEventHandler(object sender, System.ComponentModel.Design.ComponentRenameEventArgs e);
-    public partial class DesignerCollection : System.Collections.ICollection, System.Collections.IEnumerable
-    {
-        public DesignerCollection(System.Collections.IList designers) { }
-        public DesignerCollection(System.ComponentModel.Design.IDesignerHost[] designers) { }
-        public int Count { get { throw null; } }
-        public virtual System.ComponentModel.Design.IDesignerHost this[int index] { get { throw null; } }
-        int System.Collections.ICollection.Count { get { throw null; } }
-        bool System.Collections.ICollection.IsSynchronized { get { throw null; } }
-        object System.Collections.ICollection.SyncRoot { get { throw null; } }
-        public System.Collections.IEnumerator GetEnumerator() { throw null; }
-        void System.Collections.ICollection.CopyTo(System.Array array, int index) { }
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() { throw null; }
-    }
-    public partial class DesignerEventArgs : System.EventArgs
-    {
-        public DesignerEventArgs(System.ComponentModel.Design.IDesignerHost host) { }
-        public System.ComponentModel.Design.IDesignerHost Designer { get { throw null; } }
-    }
-    public delegate void DesignerEventHandler(object sender, System.ComponentModel.Design.DesignerEventArgs e);
-    public abstract partial class DesignerOptionService : System.ComponentModel.Design.IDesignerOptionService
-    {
-        protected DesignerOptionService() { }
-        public System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection Options { get { throw null; } }
-        protected System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection CreateOptionCollection(System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection parent, string name, object value) { throw null; }
-        protected virtual void PopulateOptionCollection(System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection options) { }
-        protected virtual bool ShowDialog(System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection options, object optionObject) { throw null; }
-        object System.ComponentModel.Design.IDesignerOptionService.GetOptionValue(string pageName, string valueName) { throw null; }
-        void System.ComponentModel.Design.IDesignerOptionService.SetOptionValue(string pageName, string valueName, object value) { }
-        public sealed partial class DesignerOptionCollection : System.Collections.ICollection, System.Collections.IEnumerable, System.Collections.IList
-        {
-            internal DesignerOptionCollection() { }
-            public int Count { get { throw null; } }
-            public System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection this[int index] { get { throw null; } }
-            public System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection this[string name] { get { throw null; } }
-            public string Name { get { throw null; } }
-            public System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection Parent { get { throw null; } }
-            public System.ComponentModel.PropertyDescriptorCollection Properties { get { throw null; } }
-            bool System.Collections.ICollection.IsSynchronized { get { throw null; } }
-            object System.Collections.ICollection.SyncRoot { get { throw null; } }
-            bool System.Collections.IList.IsFixedSize { get { throw null; } }
-            bool System.Collections.IList.IsReadOnly { get { throw null; } }
-            object System.Collections.IList.this[int index] { get { throw null; } set { } }
-            public void CopyTo(System.Array array, int index) { }
-            public System.Collections.IEnumerator GetEnumerator() { throw null; }
-            public int IndexOf(System.ComponentModel.Design.DesignerOptionService.DesignerOptionCollection value) { throw null; }
-            public bool ShowDialog() { throw null; }
-            int System.Collections.IList.Add(object value) { throw null; }
-            void System.Collections.IList.Clear() { }
-            bool System.Collections.IList.Contains(object value) { throw null; }
-            int System.Collections.IList.IndexOf(object value) { throw null; }
-            void System.Collections.IList.Insert(int index, object value) { }
-            void System.Collections.IList.Remove(object value) { }
-            void System.Collections.IList.RemoveAt(int index) { }
-        }
-    }
-    public abstract partial class DesignerTransaction : System.IDisposable
-    {
-        protected DesignerTransaction() { }
-        protected DesignerTransaction(string description) { }
-        public bool Canceled { get { throw null; } }
-        public bool Committed { get { throw null; } }
-        public string Description { get { throw null; } }
-        public void Cancel() { }
-        public void Commit() { }
-        protected virtual void Dispose(bool disposing) { }
-        ~DesignerTransaction() { }
-        protected abstract void OnCancel();
-        protected abstract void OnCommit();
-        void System.IDisposable.Dispose() { }
-    }
-    public partial class DesignerTransactionCloseEventArgs : System.EventArgs
-    {
-        [System.ObsoleteAttribute("This constructor is obsolete. Use DesignerTransactionCloseEventArgs(bool, bool) instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-        public DesignerTransactionCloseEventArgs(bool commit) { }
-        public DesignerTransactionCloseEventArgs(bool commit, bool lastTransaction) { }
-        public bool LastTransaction { get { throw null; } }
-        public bool TransactionCommitted { get { throw null; } }
-    }
-    public delegate void DesignerTransactionCloseEventHandler(object sender, System.ComponentModel.Design.DesignerTransactionCloseEventArgs e);
-    public partial class DesignerVerb : System.ComponentModel.Design.MenuCommand
-    {
-        public DesignerVerb(string text, System.EventHandler handler) : base(default(System.EventHandler), default(System.ComponentModel.Design.CommandID)) { }
-        public DesignerVerb(string text, System.EventHandler handler, System.ComponentModel.Design.CommandID startCommandID) : base(default(System.EventHandler), default(System.ComponentModel.Design.CommandID)) { }
-        public string Description { get { throw null; } set { } }
-        public string Text { get { throw null; } }
-        public override string ToString() { throw null; }
-    }
-    public partial class DesignerVerbCollection : System.Collections.CollectionBase
-    {
-        public DesignerVerbCollection() { }
-        public DesignerVerbCollection(System.ComponentModel.Design.DesignerVerb[] value) { }
-        public System.ComponentModel.Design.DesignerVerb this[int index] { get { throw null; } set { } }
-        public int Add(System.ComponentModel.Design.DesignerVerb value) { throw null; }
-        public void AddRange(System.ComponentModel.Design.DesignerVerb[] value) { }
-        public void AddRange(System.ComponentModel.Design.DesignerVerbCollection value) { }
-        public bool Contains(System.ComponentModel.Design.DesignerVerb value) { throw null; }
-        public void CopyTo(System.ComponentModel.Design.DesignerVerb[] array, int index) { }
-        public int IndexOf(System.ComponentModel.Design.DesignerVerb value) { throw null; }
-        public void Insert(int index, System.ComponentModel.Design.DesignerVerb value) { }
-        protected override void OnClear() { }
-        protected override void OnInsert(int index, object value) { }
-        protected override void OnRemove(int index, object value) { }
-        protected override void OnSet(int index, object oldValue, object newValue) { }
-        protected override void OnValidate(object value) { }
-        public void Remove(System.ComponentModel.Design.DesignerVerb value) { }
-    }
-    public partial class DesigntimeLicenseContext : System.ComponentModel.LicenseContext
-    {
-        public DesigntimeLicenseContext() { }
-        public override System.ComponentModel.LicenseUsageMode UsageMode { get { throw null; } }
-        public override string GetSavedLicenseKey(System.Type type, System.Reflection.Assembly resourceAssembly) { throw null; }
-        public override void SetSavedLicenseKey(System.Type type, string key) { }
-    }
-    public partial class DesigntimeLicenseContextSerializer
-    {
-        internal DesigntimeLicenseContextSerializer() { }
-        public static void Serialize(System.IO.Stream o, string cryptoKey, System.ComponentModel.Design.DesigntimeLicenseContext context) { }
-    }
-    public enum HelpContextType
-    {
-        Ambient = 0,
-        Selection = 2,
-        ToolWindowSelection = 3,
-        Window = 1,
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(32767), AllowMultiple = false, Inherited = false)]
-    public sealed partial class HelpKeywordAttribute : System.Attribute
-    {
-        public static readonly System.ComponentModel.Design.HelpKeywordAttribute Default;
-        public HelpKeywordAttribute() { }
-        public HelpKeywordAttribute(string keyword) { }
-        public HelpKeywordAttribute(System.Type t) { }
-        public string HelpKeyword { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public override bool IsDefaultAttribute() { throw null; }
-    }
-    public enum HelpKeywordType
-    {
-        F1Keyword = 0,
-        FilterKeyword = 2,
-        GeneralKeyword = 1,
-    }
-    public partial interface IComponentChangeService
-    {
-        event System.ComponentModel.Design.ComponentEventHandler ComponentAdded;
-        event System.ComponentModel.Design.ComponentEventHandler ComponentAdding;
-        event System.ComponentModel.Design.ComponentChangedEventHandler ComponentChanged;
-        event System.ComponentModel.Design.ComponentChangingEventHandler ComponentChanging;
-        event System.ComponentModel.Design.ComponentEventHandler ComponentRemoved;
-        event System.ComponentModel.Design.ComponentEventHandler ComponentRemoving;
-        event System.ComponentModel.Design.ComponentRenameEventHandler ComponentRename;
-        void OnComponentChanged(object component, System.ComponentModel.MemberDescriptor member, object oldValue, object newValue);
-        void OnComponentChanging(object component, System.ComponentModel.MemberDescriptor member);
-    }
-    public partial interface IComponentDiscoveryService
-    {
-        System.Collections.ICollection GetComponentTypes(System.ComponentModel.Design.IDesignerHost designerHost, System.Type baseType);
-    }
-    public partial interface IComponentInitializer
-    {
-        void InitializeExistingComponent(System.Collections.IDictionary defaultValues);
-        void InitializeNewComponent(System.Collections.IDictionary defaultValues);
-    }
-    public partial interface IDesigner : System.IDisposable
-    {
-        System.ComponentModel.IComponent Component { get; }
-        System.ComponentModel.Design.DesignerVerbCollection Verbs { get; }
-        void DoDefaultAction();
-        void Initialize(System.ComponentModel.IComponent component);
-    }
-    public partial interface IDesignerEventService
-    {
-        System.ComponentModel.Design.IDesignerHost ActiveDesigner { get; }
-        System.ComponentModel.Design.DesignerCollection Designers { get; }
-        event System.ComponentModel.Design.ActiveDesignerEventHandler ActiveDesignerChanged;
-        event System.ComponentModel.Design.DesignerEventHandler DesignerCreated;
-        event System.ComponentModel.Design.DesignerEventHandler DesignerDisposed;
-        event System.EventHandler SelectionChanged;
-    }
-    public partial interface IDesignerFilter
-    {
-        void PostFilterAttributes(System.Collections.IDictionary attributes);
-        void PostFilterEvents(System.Collections.IDictionary events);
-        void PostFilterProperties(System.Collections.IDictionary properties);
-        void PreFilterAttributes(System.Collections.IDictionary attributes);
-        void PreFilterEvents(System.Collections.IDictionary events);
-        void PreFilterProperties(System.Collections.IDictionary properties);
-    }
-    public partial interface IDesignerHost : System.ComponentModel.Design.IServiceContainer, System.IServiceProvider
-    {
-        System.ComponentModel.IContainer Container { get; }
-        bool InTransaction { get; }
-        bool Loading { get; }
-        System.ComponentModel.IComponent RootComponent { get; }
-        string RootComponentClassName { get; }
-        string TransactionDescription { get; }
-        event System.EventHandler Activated;
-        event System.EventHandler Deactivated;
-        event System.EventHandler LoadComplete;
-        event System.ComponentModel.Design.DesignerTransactionCloseEventHandler TransactionClosed;
-        event System.ComponentModel.Design.DesignerTransactionCloseEventHandler TransactionClosing;
-        event System.EventHandler TransactionOpened;
-        event System.EventHandler TransactionOpening;
-        void Activate();
-        System.ComponentModel.IComponent CreateComponent(System.Type componentClass);
-        System.ComponentModel.IComponent CreateComponent(System.Type componentClass, string name);
-        System.ComponentModel.Design.DesignerTransaction CreateTransaction();
-        System.ComponentModel.Design.DesignerTransaction CreateTransaction(string description);
-        void DestroyComponent(System.ComponentModel.IComponent component);
-        System.ComponentModel.Design.IDesigner GetDesigner(System.ComponentModel.IComponent component);
-        System.Type GetType(string typeName);
-    }
-    public partial interface IDesignerHostTransactionState
-    {
-        bool IsClosingTransaction { get; }
-    }
-    public partial interface IDesignerOptionService
-    {
-        object GetOptionValue(string pageName, string valueName);
-        void SetOptionValue(string pageName, string valueName, object value);
-    }
-    public partial interface IDictionaryService
-    {
-        object GetKey(object value);
-        object GetValue(object key);
-        void SetValue(object key, object value);
-    }
-    public partial interface IEventBindingService
-    {
-        string CreateUniqueMethodName(System.ComponentModel.IComponent component, System.ComponentModel.EventDescriptor e);
-        System.Collections.ICollection GetCompatibleMethods(System.ComponentModel.EventDescriptor e);
-        System.ComponentModel.EventDescriptor GetEvent(System.ComponentModel.PropertyDescriptor property);
-        System.ComponentModel.PropertyDescriptorCollection GetEventProperties(System.ComponentModel.EventDescriptorCollection events);
-        System.ComponentModel.PropertyDescriptor GetEventProperty(System.ComponentModel.EventDescriptor e);
-        bool ShowCode();
-        bool ShowCode(System.ComponentModel.IComponent component, System.ComponentModel.EventDescriptor e);
-        bool ShowCode(int lineNumber);
-    }
-    public partial interface IExtenderListService
-    {
-        System.ComponentModel.IExtenderProvider[] GetExtenderProviders();
-    }
-    public partial interface IExtenderProviderService
-    {
-        void AddExtenderProvider(System.ComponentModel.IExtenderProvider provider);
-        void RemoveExtenderProvider(System.ComponentModel.IExtenderProvider provider);
-    }
-    public partial interface IHelpService
-    {
-        void AddContextAttribute(string name, string value, System.ComponentModel.Design.HelpKeywordType keywordType);
-        void ClearContextAttributes();
-        System.ComponentModel.Design.IHelpService CreateLocalContext(System.ComponentModel.Design.HelpContextType contextType);
-        void RemoveContextAttribute(string name, string value);
-        void RemoveLocalContext(System.ComponentModel.Design.IHelpService localContext);
-        void ShowHelpFromKeyword(string helpKeyword);
-        void ShowHelpFromUrl(string helpUrl);
-    }
-    public partial interface IInheritanceService
-    {
-        void AddInheritedComponents(System.ComponentModel.IComponent component, System.ComponentModel.IContainer container);
-        System.ComponentModel.InheritanceAttribute GetInheritanceAttribute(System.ComponentModel.IComponent component);
-    }
-    public partial interface IMenuCommandService
-    {
-        System.ComponentModel.Design.DesignerVerbCollection Verbs { get; }
-        void AddCommand(System.ComponentModel.Design.MenuCommand command);
-        void AddVerb(System.ComponentModel.Design.DesignerVerb verb);
-        System.ComponentModel.Design.MenuCommand FindCommand(System.ComponentModel.Design.CommandID commandID);
-        bool GlobalInvoke(System.ComponentModel.Design.CommandID commandID);
-        void RemoveCommand(System.ComponentModel.Design.MenuCommand command);
-        void RemoveVerb(System.ComponentModel.Design.DesignerVerb verb);
-        void ShowContextMenu(System.ComponentModel.Design.CommandID menuID, int x, int y);
-    }
-    public partial interface IReferenceService
-    {
-        System.ComponentModel.IComponent GetComponent(object reference);
-        string GetName(object reference);
-        object GetReference(string name);
-        object[] GetReferences();
-        object[] GetReferences(System.Type baseType);
-    }
-    public partial interface IResourceService
-    {
-        System.Resources.IResourceReader GetResourceReader(System.Globalization.CultureInfo info);
-        System.Resources.IResourceWriter GetResourceWriter(System.Globalization.CultureInfo info);
-    }
-    public partial interface IRootDesigner : System.ComponentModel.Design.IDesigner, System.IDisposable
-    {
-        System.ComponentModel.Design.ViewTechnology[] SupportedTechnologies { get; }
-        object GetView(System.ComponentModel.Design.ViewTechnology technology);
-    }
-    public partial interface ISelectionService
-    {
-        object PrimarySelection { get; }
-        int SelectionCount { get; }
-        event System.EventHandler SelectionChanged;
-        event System.EventHandler SelectionChanging;
-        bool GetComponentSelected(object component);
-        System.Collections.ICollection GetSelectedComponents();
-        void SetSelectedComponents(System.Collections.ICollection components);
-        void SetSelectedComponents(System.Collections.ICollection components, System.ComponentModel.Design.SelectionTypes selectionType);
-    }
-    public partial interface IServiceContainer : System.IServiceProvider
-    {
-        void AddService(System.Type serviceType, System.ComponentModel.Design.ServiceCreatorCallback callback);
-        void AddService(System.Type serviceType, System.ComponentModel.Design.ServiceCreatorCallback callback, bool promote);
-        void AddService(System.Type serviceType, object serviceInstance);
-        void AddService(System.Type serviceType, object serviceInstance, bool promote);
-        void RemoveService(System.Type serviceType);
-        void RemoveService(System.Type serviceType, bool promote);
-    }
-    public partial interface ITreeDesigner : System.ComponentModel.Design.IDesigner, System.IDisposable
-    {
-        System.Collections.ICollection Children { get; }
-        System.ComponentModel.Design.IDesigner Parent { get; }
-    }
-    public partial interface ITypeDescriptorFilterService
-    {
-        bool FilterAttributes(System.ComponentModel.IComponent component, System.Collections.IDictionary attributes);
-        bool FilterEvents(System.ComponentModel.IComponent component, System.Collections.IDictionary events);
-        bool FilterProperties(System.ComponentModel.IComponent component, System.Collections.IDictionary properties);
-    }
-    public partial interface ITypeDiscoveryService
-    {
-        System.Collections.ICollection GetTypes(System.Type baseType, bool excludeGlobalTypes);
-    }
-    public partial interface ITypeResolutionService
-    {
-        System.Reflection.Assembly GetAssembly(System.Reflection.AssemblyName name);
-        System.Reflection.Assembly GetAssembly(System.Reflection.AssemblyName name, bool throwOnError);
-        string GetPathOfAssembly(System.Reflection.AssemblyName name);
-        System.Type GetType(string name);
-        System.Type GetType(string name, bool throwOnError);
-        System.Type GetType(string name, bool throwOnError, bool ignoreCase);
-        void ReferenceAssembly(System.Reflection.AssemblyName name);
-    }
-    public partial class MenuCommand
-    {
-        public MenuCommand(System.EventHandler handler, System.ComponentModel.Design.CommandID command) { }
-        public virtual bool Checked { get { throw null; } set { } }
-        public virtual System.ComponentModel.Design.CommandID CommandID { get { throw null; } }
-        public virtual bool Enabled { get { throw null; } set { } }
-        public virtual int OleStatus { get { throw null; } }
-        public virtual System.Collections.IDictionary Properties { get { throw null; } }
-        public virtual bool Supported { get { throw null; } set { } }
-        public virtual bool Visible { get { throw null; } set { } }
-        public event System.EventHandler CommandChanged { add { } remove { } }
-        public virtual void Invoke() { }
-        public virtual void Invoke(object arg) { }
-        protected virtual void OnCommandChanged(System.EventArgs e) { }
-        public override string ToString() { throw null; }
-    }
-    [System.FlagsAttribute]
-    public enum SelectionTypes
-    {
-        Add = 64,
-        Auto = 1,
-        [System.ObsoleteAttribute("This value has been deprecated. Use SelectionTypes.Primary instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-        Click = 16,
-        [System.ObsoleteAttribute("This value has been deprecated.  It is no longer supported. https://go.microsoft.com/fwlink/?linkid=14202")]
-        MouseDown = 4,
-        [System.ObsoleteAttribute("This value has been deprecated.  It is no longer supported. https://go.microsoft.com/fwlink/?linkid=14202")]
-        MouseUp = 8,
-        [System.ObsoleteAttribute("This value has been deprecated. Use SelectionTypes.Auto instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-        Normal = 1,
-        Primary = 16,
-        Remove = 128,
-        Replace = 2,
-        Toggle = 32,
-        [System.ObsoleteAttribute("This value has been deprecated. Use Enum class methods to determine valid values, or use a type converter. https://go.microsoft.com/fwlink/?linkid=14202")]
-        Valid = 31,
-    }
-    public partial class ServiceContainer : System.ComponentModel.Design.IServiceContainer, System.IDisposable, System.IServiceProvider
-    {
-        public ServiceContainer() { }
-        public ServiceContainer(System.IServiceProvider parentProvider) { }
-        protected virtual System.Type[] DefaultServices { get { throw null; } }
-        public void AddService(System.Type serviceType, System.ComponentModel.Design.ServiceCreatorCallback callback) { }
-        public virtual void AddService(System.Type serviceType, System.ComponentModel.Design.ServiceCreatorCallback callback, bool promote) { }
-        public void AddService(System.Type serviceType, object serviceInstance) { }
-        public virtual void AddService(System.Type serviceType, object serviceInstance, bool promote) { }
-        public void Dispose() { }
-        protected virtual void Dispose(bool disposing) { }
-        public virtual object GetService(System.Type serviceType) { throw null; }
-        public void RemoveService(System.Type serviceType) { }
-        public virtual void RemoveService(System.Type serviceType, bool promote) { }
-    }
-    public delegate object ServiceCreatorCallback(System.ComponentModel.Design.IServiceContainer container, System.Type serviceType);
-    public partial class StandardCommands
-    {
-        public static readonly System.ComponentModel.Design.CommandID AlignBottom;
-        public static readonly System.ComponentModel.Design.CommandID AlignHorizontalCenters;
-        public static readonly System.ComponentModel.Design.CommandID AlignLeft;
-        public static readonly System.ComponentModel.Design.CommandID AlignRight;
-        public static readonly System.ComponentModel.Design.CommandID AlignToGrid;
-        public static readonly System.ComponentModel.Design.CommandID AlignTop;
-        public static readonly System.ComponentModel.Design.CommandID AlignVerticalCenters;
-        public static readonly System.ComponentModel.Design.CommandID ArrangeBottom;
-        public static readonly System.ComponentModel.Design.CommandID ArrangeIcons;
-        public static readonly System.ComponentModel.Design.CommandID ArrangeRight;
-        public static readonly System.ComponentModel.Design.CommandID BringForward;
-        public static readonly System.ComponentModel.Design.CommandID BringToFront;
-        public static readonly System.ComponentModel.Design.CommandID CenterHorizontally;
-        public static readonly System.ComponentModel.Design.CommandID CenterVertically;
-        public static readonly System.ComponentModel.Design.CommandID Copy;
-        public static readonly System.ComponentModel.Design.CommandID Cut;
-        public static readonly System.ComponentModel.Design.CommandID Delete;
-        public static readonly System.ComponentModel.Design.CommandID DocumentOutline;
-        public static readonly System.ComponentModel.Design.CommandID F1Help;
-        public static readonly System.ComponentModel.Design.CommandID Group;
-        public static readonly System.ComponentModel.Design.CommandID HorizSpaceConcatenate;
-        public static readonly System.ComponentModel.Design.CommandID HorizSpaceDecrease;
-        public static readonly System.ComponentModel.Design.CommandID HorizSpaceIncrease;
-        public static readonly System.ComponentModel.Design.CommandID HorizSpaceMakeEqual;
-        public static readonly System.ComponentModel.Design.CommandID LineupIcons;
-        public static readonly System.ComponentModel.Design.CommandID LockControls;
-        public static readonly System.ComponentModel.Design.CommandID MultiLevelRedo;
-        public static readonly System.ComponentModel.Design.CommandID MultiLevelUndo;
-        public static readonly System.ComponentModel.Design.CommandID Paste;
-        public static readonly System.ComponentModel.Design.CommandID Properties;
-        public static readonly System.ComponentModel.Design.CommandID PropertiesWindow;
-        public static readonly System.ComponentModel.Design.CommandID Redo;
-        public static readonly System.ComponentModel.Design.CommandID Replace;
-        public static readonly System.ComponentModel.Design.CommandID SelectAll;
-        public static readonly System.ComponentModel.Design.CommandID SendBackward;
-        public static readonly System.ComponentModel.Design.CommandID SendToBack;
-        public static readonly System.ComponentModel.Design.CommandID ShowGrid;
-        public static readonly System.ComponentModel.Design.CommandID ShowLargeIcons;
-        public static readonly System.ComponentModel.Design.CommandID SizeToControl;
-        public static readonly System.ComponentModel.Design.CommandID SizeToControlHeight;
-        public static readonly System.ComponentModel.Design.CommandID SizeToControlWidth;
-        public static readonly System.ComponentModel.Design.CommandID SizeToFit;
-        public static readonly System.ComponentModel.Design.CommandID SizeToGrid;
-        public static readonly System.ComponentModel.Design.CommandID SnapToGrid;
-        public static readonly System.ComponentModel.Design.CommandID TabOrder;
-        public static readonly System.ComponentModel.Design.CommandID Undo;
-        public static readonly System.ComponentModel.Design.CommandID Ungroup;
-        public static readonly System.ComponentModel.Design.CommandID VerbFirst;
-        public static readonly System.ComponentModel.Design.CommandID VerbLast;
-        public static readonly System.ComponentModel.Design.CommandID VertSpaceConcatenate;
-        public static readonly System.ComponentModel.Design.CommandID VertSpaceDecrease;
-        public static readonly System.ComponentModel.Design.CommandID VertSpaceIncrease;
-        public static readonly System.ComponentModel.Design.CommandID VertSpaceMakeEqual;
-        public static readonly System.ComponentModel.Design.CommandID ViewCode;
-        public static readonly System.ComponentModel.Design.CommandID ViewGrid;
-        public StandardCommands() { }
-    }
-    public partial class StandardToolWindows
-    {
-        public static readonly System.Guid ObjectBrowser;
-        public static readonly System.Guid OutputWindow;
-        public static readonly System.Guid ProjectExplorer;
-        public static readonly System.Guid PropertyBrowser;
-        public static readonly System.Guid RelatedLinks;
-        public static readonly System.Guid ServerExplorer;
-        public static readonly System.Guid TaskList;
-        public static readonly System.Guid Toolbox;
-        public StandardToolWindows() { }
-    }
-    public abstract partial class TypeDescriptionProviderService
-    {
-        protected TypeDescriptionProviderService() { }
-        public abstract System.ComponentModel.TypeDescriptionProvider GetProvider(object instance);
-        public abstract System.ComponentModel.TypeDescriptionProvider GetProvider(System.Type type);
-    }
-    public enum ViewTechnology
-    {
-        Default = 2,
-        [System.ObsoleteAttribute("This value has been deprecated. Use ViewTechnology.Default instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-        Passthrough = 0,
-        [System.ObsoleteAttribute("This value has been deprecated. Use ViewTechnology.Default instead.  https://go.microsoft.com/fwlink/?linkid=14202")]
-        WindowsForms = 1,
-    }
-}
-namespace System.ComponentModel.Design.Serialization
-{
-    public abstract partial class ComponentSerializationService
-    {
-        protected ComponentSerializationService() { }
-        public abstract System.ComponentModel.Design.Serialization.SerializationStore CreateStore();
-        public abstract System.Collections.ICollection Deserialize(System.ComponentModel.Design.Serialization.SerializationStore store);
-        public abstract System.Collections.ICollection Deserialize(System.ComponentModel.Design.Serialization.SerializationStore store, System.ComponentModel.IContainer container);
-        public void DeserializeTo(System.ComponentModel.Design.Serialization.SerializationStore store, System.ComponentModel.IContainer container) { }
-        public void DeserializeTo(System.ComponentModel.Design.Serialization.SerializationStore store, System.ComponentModel.IContainer container, bool validateRecycledTypes) { }
-        public abstract void DeserializeTo(System.ComponentModel.Design.Serialization.SerializationStore store, System.ComponentModel.IContainer container, bool validateRecycledTypes, bool applyDefaults);
-        public abstract System.ComponentModel.Design.Serialization.SerializationStore LoadStore(System.IO.Stream stream);
-        public abstract void Serialize(System.ComponentModel.Design.Serialization.SerializationStore store, object value);
-        public abstract void SerializeAbsolute(System.ComponentModel.Design.Serialization.SerializationStore store, object value);
-        public abstract void SerializeMember(System.ComponentModel.Design.Serialization.SerializationStore store, object owningObject, System.ComponentModel.MemberDescriptor member);
-        public abstract void SerializeMemberAbsolute(System.ComponentModel.Design.Serialization.SerializationStore store, object owningObject, System.ComponentModel.MemberDescriptor member);
-    }
-    public sealed partial class ContextStack
-    {
-        public ContextStack() { }
-        public object Current { get { throw null; } }
-        public object this[int level] { get { throw null; } }
-        public object this[System.Type type] { get { throw null; } }
-        public void Append(object context) { }
-        public object Pop() { throw null; }
-        public void Push(object context) { }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(4), Inherited = false)]
-    public sealed partial class DefaultSerializationProviderAttribute : System.Attribute
-    {
-        public DefaultSerializationProviderAttribute(string providerTypeName) { }
-        public DefaultSerializationProviderAttribute(System.Type providerType) { }
-        public string ProviderTypeName { get { throw null; } }
-    }
-    public abstract partial class DesignerLoader
-    {
-        protected DesignerLoader() { }
-        public virtual bool Loading { get { throw null; } }
-        public abstract void BeginLoad(System.ComponentModel.Design.Serialization.IDesignerLoaderHost host);
-        public abstract void Dispose();
-        public virtual void Flush() { }
-    }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(1028), AllowMultiple = true, Inherited = true)]
-    public sealed partial class DesignerSerializerAttribute : System.Attribute
-    {
-        public DesignerSerializerAttribute(string serializerTypeName, string baseSerializerTypeName) { }
-        public DesignerSerializerAttribute(string serializerTypeName, System.Type baseSerializerType) { }
-        public DesignerSerializerAttribute(System.Type serializerType, System.Type baseSerializerType) { }
-        public string SerializerBaseTypeName { get { throw null; } }
-        public string SerializerTypeName { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-    }
-    public partial interface IDesignerLoaderHost : System.ComponentModel.Design.IDesignerHost, System.ComponentModel.Design.IServiceContainer, System.IServiceProvider
-    {
-        void EndLoad(string baseClassName, bool successful, System.Collections.ICollection errorCollection);
-        void Reload();
-    }
-    public partial interface IDesignerLoaderHost2 : System.ComponentModel.Design.IDesignerHost, System.ComponentModel.Design.IServiceContainer, System.ComponentModel.Design.Serialization.IDesignerLoaderHost, System.IServiceProvider
-    {
-        bool CanReloadWithErrors { get; set; }
-        bool IgnoreErrorsDuringReload { get; set; }
-    }
-    public partial interface IDesignerLoaderService
-    {
-        void AddLoadDependency();
-        void DependentLoadComplete(bool successful, System.Collections.ICollection errorCollection);
-        bool Reload();
-    }
-    public partial interface IDesignerSerializationManager : System.IServiceProvider
-    {
-        System.ComponentModel.Design.Serialization.ContextStack Context { get; }
-        System.ComponentModel.PropertyDescriptorCollection Properties { get; }
-        event System.ComponentModel.Design.Serialization.ResolveNameEventHandler ResolveName;
-        event System.EventHandler SerializationComplete;
-        void AddSerializationProvider(System.ComponentModel.Design.Serialization.IDesignerSerializationProvider provider);
-        object CreateInstance(System.Type type, System.Collections.ICollection arguments, string name, bool addToContainer);
-        object GetInstance(string name);
-        string GetName(object value);
-        object GetSerializer(System.Type objectType, System.Type serializerType);
-        System.Type GetType(string typeName);
-        void RemoveSerializationProvider(System.ComponentModel.Design.Serialization.IDesignerSerializationProvider provider);
-        void ReportError(object errorInformation);
-        void SetName(object instance, string name);
-    }
-    public partial interface IDesignerSerializationProvider
-    {
-        object GetSerializer(System.ComponentModel.Design.Serialization.IDesignerSerializationManager manager, object currentSerializer, System.Type objectType, System.Type serializerType);
-    }
-    public partial interface IDesignerSerializationService
-    {
-        System.Collections.ICollection Deserialize(object serializationData);
-        object Serialize(System.Collections.ICollection objects);
-    }
-    public partial interface INameCreationService
-    {
-        string CreateName(System.ComponentModel.IContainer container, System.Type dataType);
-        bool IsValidName(string name);
-        void ValidateName(string name);
-    }
-    public sealed partial class InstanceDescriptor
-    {
-        public InstanceDescriptor(System.Reflection.MemberInfo member, System.Collections.ICollection arguments) { }
-        public InstanceDescriptor(System.Reflection.MemberInfo member, System.Collections.ICollection arguments, bool isComplete) { }
-        public System.Collections.ICollection Arguments { get { throw null; } }
-        public bool IsComplete { get { throw null; } }
-        public System.Reflection.MemberInfo MemberInfo { get { throw null; } }
-        public object Invoke() { throw null; }
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public readonly partial struct MemberRelationship
-    {
-        private readonly object _dummy;
-        public static readonly System.ComponentModel.Design.Serialization.MemberRelationship Empty;
-        public MemberRelationship(object owner, System.ComponentModel.MemberDescriptor member) { throw null; }
-        public bool IsEmpty { get { throw null; } }
-        public System.ComponentModel.MemberDescriptor Member { get { throw null; } }
-        public object Owner { get { throw null; } }
-        public override bool Equals(object obj) { throw null; }
-        public override int GetHashCode() { throw null; }
-        public static bool operator ==(System.ComponentModel.Design.Serialization.MemberRelationship left, System.ComponentModel.Design.Serialization.MemberRelationship right) { throw null; }
-        public static bool operator !=(System.ComponentModel.Design.Serialization.MemberRelationship left, System.ComponentModel.Design.Serialization.MemberRelationship right) { throw null; }
-    }
-    public abstract partial class MemberRelationshipService
-    {
-        protected MemberRelationshipService() { }
-        public System.ComponentModel.Design.Serialization.MemberRelationship this[System.ComponentModel.Design.Serialization.MemberRelationship source] { get { throw null; } set { } }
-        public System.ComponentModel.Design.Serialization.MemberRelationship this[object sourceOwner, System.ComponentModel.MemberDescriptor sourceMember] { get { throw null; } set { } }
-        protected virtual System.ComponentModel.Design.Serialization.MemberRelationship GetRelationship(System.ComponentModel.Design.Serialization.MemberRelationship source) { throw null; }
-        protected virtual void SetRelationship(System.ComponentModel.Design.Serialization.MemberRelationship source, System.ComponentModel.Design.Serialization.MemberRelationship relationship) { }
-        public abstract bool SupportsRelationship(System.ComponentModel.Design.Serialization.MemberRelationship source, System.ComponentModel.Design.Serialization.MemberRelationship relationship);
-    }
-    public partial class ResolveNameEventArgs : System.EventArgs
-    {
-        public ResolveNameEventArgs(string name) { }
-        public string Name { get { throw null; } }
-        public object Value { get { throw null; } set { } }
-    }
-    public delegate void ResolveNameEventHandler(object sender, System.ComponentModel.Design.Serialization.ResolveNameEventArgs e);
-    [System.AttributeUsageAttribute((System.AttributeTargets)(1028), AllowMultiple = true, Inherited = true)]
-    [System.ObsoleteAttribute("This attribute has been deprecated. Use DesignerSerializerAttribute instead.  For example, to specify a root designer for CodeDom, use DesignerSerializerAttribute(...,typeof(TypeCodeDomSerializer)).  https://go.microsoft.com/fwlink/?linkid=14202")]
-    public sealed partial class RootDesignerSerializerAttribute : System.Attribute
-    {
-        public RootDesignerSerializerAttribute(string serializerTypeName, string baseSerializerTypeName, bool reloadable) { }
-        public RootDesignerSerializerAttribute(string serializerTypeName, System.Type baseSerializerType, bool reloadable) { }
-        public RootDesignerSerializerAttribute(System.Type serializerType, System.Type baseSerializerType, bool reloadable) { }
-        public bool Reloadable { get { throw null; } }
-        public string SerializerBaseTypeName { get { throw null; } }
-        public string SerializerTypeName { get { throw null; } }
-        public override object TypeId { get { throw null; } }
-    }
-    public abstract partial class SerializationStore : System.IDisposable
-    {
-        protected SerializationStore() { }
-        public abstract System.Collections.ICollection Errors { get; }
-        public abstract void Close();
-        protected virtual void Dispose(bool disposing) { }
-        public abstract void Save(System.IO.Stream stream);
-        void System.IDisposable.Dispose() { }
-    }
-}
