+++ src/System.Data.SqlClient/ref/System.Data.SqlClient.Manual.cs
@@ -10,12 +10,12 @@ namespace System.Data.SqlClient
     public sealed partial class SqlCommand : System.Data.Common.DbCommand
     {
         // SqlCommand expects IDisposable methods to be implemented via System.ComponentModel.Component, which it no longer inherits from
-        override protected void Dispose(bool disposing) { }
+        protected override void Dispose(bool disposing) { }
     }
     public sealed partial class SqlConnection : System.Data.Common.DbConnection
     {
         // SqlConection expects IDisposable methods to be implemented via System.ComponentModel.Component, which it no longer inherits from
-        override protected void Dispose(bool disposing) { }
+        protected override void Dispose(bool disposing) { }
     }
     [System.ComponentModel.TypeConverter(typeof(SqlParameterConverter))]
     public sealed partial class SqlParameter
