+++ src/System.Data.SqlClient/ref/System.Data.SqlClient.NetCoreApp.cs
@@ -7,7 +7,7 @@
 
 namespace Microsoft.SqlServer.Server
 {
-    [System.AttributeUsageAttribute(System.AttributeTargets.Field | System.AttributeTargets.Property | System.AttributeTargets.ReturnValue | System.AttributeTargets.Parameter, AllowMultiple=false, Inherited=false)]
+    [System.AttributeUsageAttribute(System.AttributeTargets.Field | System.AttributeTargets.Parameter | System.AttributeTargets.Property | System.AttributeTargets.ReturnValue, AllowMultiple=false, Inherited=false)]
     public partial class SqlFacetAttribute : System.Attribute
     {
         public SqlFacetAttribute() { }
@@ -24,38 +24,34 @@ namespace System.Data.SqlClient
     {
         public System.Collections.ObjectModel.ReadOnlyCollection<System.Data.Common.DbColumn> GetColumnSchema() { throw null; }
     }
-
     public enum PoolBlockingPeriod
     {
-        Auto = 0,
         AlwaysBlock = 1,
+        Auto = 0,
         NeverBlock = 2,
     }
-    
     public sealed partial class SqlConnectionStringBuilder : System.Data.Common.DbConnectionStringBuilder
     {
-        public PoolBlockingPeriod PoolBlockingPeriod { get { throw null; } set { } }
+        public System.Data.SqlClient.PoolBlockingPeriod PoolBlockingPeriod { get { throw null; } set { } }
     }
-
 }
-
 namespace System.Data.SqlTypes
 {
     public sealed partial class SqlFileStream : System.IO.Stream
     {
         public SqlFileStream(string path, byte[] transactionContext, System.IO.FileAccess access) { }
-        public SqlFileStream(string path, byte[] transactionContext, System.IO.FileAccess access, System.IO.FileOptions options, Int64 allocationSize) { }
-        public string Name { get { throw null; } }
-        public byte[] TransactionContext { get { throw null; } }
+        public SqlFileStream(string path, byte[] transactionContext, System.IO.FileAccess access, System.IO.FileOptions options, long allocationSize) { }
         public override bool CanRead { get { throw null; } }
         public override bool CanSeek { get { throw null; } }
         public override bool CanWrite { get { throw null; } }
         public override long Length { get { throw null; } }
-        public override long Position { get { throw null; } set { throw null; } }
+        public string Name { get { throw null; } }
+        public override long Position { get { throw null; } set { } }
+        public byte[] TransactionContext { get { throw null; } }
         public override void Flush() { }
         public override int Read(byte[] buffer, int offset, int count) { throw null; }
         public override long Seek(long offset, System.IO.SeekOrigin origin) { throw null; }
-        public override void SetLength(long value) { throw null; }
-        public override void Write(byte[] buffer, int offset, int count) { throw null; }
+        public override void SetLength(long value) { }
+        public override void Write(byte[] buffer, int offset, int count) { }
     }
 }
