+++ src/System.Diagnostics.DiagnosticSource/ref/System.Diagnostics.DiagnosticSource.cs
@@ -7,28 +7,6 @@
 
 namespace System.Diagnostics
 {
-    public partial class Activity
-    {
-        public Activity(string operationName) { }
-        public System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string, string>> Baggage { get { throw null; } }
-        public static System.Diagnostics.Activity Current { get { throw null; } set { } }
-        public System.TimeSpan Duration { get { throw null; } }
-        public string Id { get { throw null; } }
-        public string OperationName { get { throw null; } }
-        public System.Diagnostics.Activity Parent { get { throw null; } }
-        public string ParentId { get { throw null; } }
-        public string RootId { get { throw null; } }
-        public System.DateTime StartTimeUtc { get { throw null; } }
-        public System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string, string>> Tags { get { throw null; } }
-        public System.Diagnostics.Activity AddBaggage(string key, string value) { throw null; }
-        public System.Diagnostics.Activity AddTag(string key, string value) { throw null; }
-        public string GetBaggageItem(string key) { throw null; }
-        public System.Diagnostics.Activity SetEndTime(System.DateTime endTimeUtc) { throw null; }
-        public System.Diagnostics.Activity SetParentId(string parentId) { throw null; }
-        public System.Diagnostics.Activity SetStartTime(System.DateTime startTimeUtc) { throw null; }
-        public System.Diagnostics.Activity Start() { throw null; }
-        public void Stop() { }
-    }
     public partial class DiagnosticListener : System.Diagnostics.DiagnosticSource, System.IDisposable, System.IObservable<System.Collections.Generic.KeyValuePair<string, object>>
     {
         public DiagnosticListener(string name) { }
@@ -49,8 +27,6 @@ namespace System.Diagnostics
         protected DiagnosticSource() { }
         public abstract bool IsEnabled(string name);
         public virtual bool IsEnabled(string name, object arg1, object arg2 = null) { throw null; }
-        public System.Diagnostics.Activity StartActivity(System.Diagnostics.Activity activity, object args) { throw null; }
-        public void StopActivity(System.Diagnostics.Activity activity, object args) { }
         public abstract void Write(string name, object value);
     }
 }
