--- src/System.Diagnostics.Process/ref/System.Diagnostics.Process.Manual.cs
@@ -1,14 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the http://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-namespace System.Diagnostics
-{
-    public sealed partial class ProcessStartInfo
-    {
-        public string PasswordInClearText { get { throw null; } set { } }
-    }
-}
