+++ src/System.IO.Packaging/ref/System.IO.Packaging.Serialization.cs
@@ -5,7 +5,6 @@
 // Changes to this file must follow the http://aka.ms/api-review process.
 // ------------------------------------------------------------------------------
 
-
 namespace System.IO
 {
     public partial class FileFormatException : System.FormatException
