+++ src/System.Net.Requests/ref/System.Net.Requests.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Net.Requests.cs" />
-    <Compile Include="System.Net.Requests.Manual.cs" />
+    <Compile Include="System.Net.Requests.Forwards.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\System.Collections.Specialized\ref\System.Collections.Specialized.csproj" />
