+++ src/System.Reflection.Emit.Lightweight/ref/System.Reflection.Emit.Lightweight.cs
@@ -7,28 +7,6 @@
 
 namespace System.Reflection.Emit
 {
-    public sealed partial class DynamicILInfo
-    {
-        internal DynamicILInfo() { }
-        public System.Reflection.Emit.DynamicMethod DynamicMethod { get { throw null; } }
-        public int GetTokenFor(byte[] signature) { throw null; }
-        public int GetTokenFor(System.Reflection.Emit.DynamicMethod method) { throw null; }
-        public int GetTokenFor(System.RuntimeFieldHandle field) { throw null; }
-        public int GetTokenFor(System.RuntimeFieldHandle field, System.RuntimeTypeHandle contextType) { throw null; }
-        public int GetTokenFor(System.RuntimeMethodHandle method) { throw null; }
-        public int GetTokenFor(System.RuntimeMethodHandle method, System.RuntimeTypeHandle contextType) { throw null; }
-        public int GetTokenFor(System.RuntimeTypeHandle type) { throw null; }
-        public int GetTokenFor(string literal) { throw null; }
-        [System.CLSCompliantAttribute(false)]
-        public unsafe void SetCode(byte* code, int codeSize, int maxStackSize) { }
-        public void SetCode(byte[] code, int maxStackSize) { }
-        [System.CLSCompliantAttribute(false)]
-        public unsafe void SetExceptions(byte* exceptions, int exceptionsSize) { }
-        public void SetExceptions(byte[] exceptions) { }
-        [System.CLSCompliantAttribute(false)]
-        public unsafe void SetLocalSignature(byte* localSignature, int signatureSize) { }
-        public void SetLocalSignature(byte[] localSignature) { }
-    }
     public sealed partial class DynamicMethod : System.Reflection.MethodInfo
     {
         public DynamicMethod(string name, System.Reflection.MethodAttributes attributes, System.Reflection.CallingConventions callingConvention, System.Type returnType, System.Type[] parameterTypes, System.Reflection.Module m, bool skipVisibility) { }
@@ -55,7 +33,6 @@ namespace System.Reflection.Emit
         public override System.Reflection.MethodInfo GetBaseDefinition() { throw null; }
         public override object[] GetCustomAttributes(bool inherit) { throw null; }
         public override object[] GetCustomAttributes(System.Type attributeType, bool inherit) { throw null; }
-        public System.Reflection.Emit.DynamicILInfo GetDynamicILInfo() { throw null; }
         public System.Reflection.Emit.ILGenerator GetILGenerator() { throw null; }
         public System.Reflection.Emit.ILGenerator GetILGenerator(int streamSize) { throw null; }
         public override System.Reflection.MethodImplAttributes GetMethodImplementationFlags() { throw null; }
