+++ src/System.Reflection.Emit.Lightweight/ref/System.Reflection.Emit.Lightweight.netcore.cs
@@ -5,7 +5,6 @@
 // Changes to this file must follow the http://aka.ms/api-review process.
 // ------------------------------------------------------------------------------
 
-
 namespace System.Reflection.Emit
 {
     public sealed partial class DynamicILInfo
@@ -13,12 +12,12 @@ namespace System.Reflection.Emit
         internal DynamicILInfo() { }
         public System.Reflection.Emit.DynamicMethod DynamicMethod { get { throw null; } }
         public int GetTokenFor(byte[] signature) { throw null; }
-        public int GetTokenFor(DynamicMethod method) { throw null; }
-        public int GetTokenFor(RuntimeFieldHandle field) { throw null; }
-        public int GetTokenFor(RuntimeFieldHandle field, RuntimeTypeHandle contextType) { throw null; }
-        public int GetTokenFor(RuntimeMethodHandle method) { throw null; }
-        public int GetTokenFor(RuntimeMethodHandle method, RuntimeTypeHandle contextType) { throw null; }
-        public int GetTokenFor(RuntimeTypeHandle type) { throw null; }
+        public int GetTokenFor(System.Reflection.Emit.DynamicMethod method) { throw null; }
+        public int GetTokenFor(System.RuntimeFieldHandle field) { throw null; }
+        public int GetTokenFor(System.RuntimeFieldHandle field, System.RuntimeTypeHandle contextType) { throw null; }
+        public int GetTokenFor(System.RuntimeMethodHandle method) { throw null; }
+        public int GetTokenFor(System.RuntimeMethodHandle method, System.RuntimeTypeHandle contextType) { throw null; }
+        public int GetTokenFor(System.RuntimeTypeHandle type) { throw null; }
         public int GetTokenFor(string literal) { throw null; }
         [System.CLSCompliantAttribute(false)]
         public unsafe void SetCode(byte* code, int codeSize, int maxStackSize) { }
@@ -30,9 +29,8 @@ namespace System.Reflection.Emit
         public unsafe void SetLocalSignature(byte* localSignature, int signatureSize) { }
         public void SetLocalSignature(byte[] localSignature) { }
     }
-
     public sealed partial class DynamicMethod : System.Reflection.MethodInfo
     {
-        public DynamicILInfo GetDynamicILInfo() { throw null; }
+        public System.Reflection.Emit.DynamicILInfo GetDynamicILInfo() { throw null; }
     }
 }
