+++ src/System.Runtime.Extensions/ref/System.Runtime.Extensions.Manual.cs
@@ -11,12 +11,12 @@ namespace System
 {
     public static partial class Math
     {
-        public const double PI = 3.14159265358979323846;
-        public const double E = 2.7182818284590452354;
+        public const double E = 2.7182818284590451;
+        public const double PI = 3.1415926535897931;
     }
     public static partial class MathF
     {
-        public const float PI = 3.14159265f;
-        public const float E = 2.71828183f;
+        public const float E = 2.71828175f;
+        public const float PI = 3.14159274f;
     }
 }
