--- src/System.Security.Cryptography.Cng/ref/System.Security.Cryptography.Cng.ECDiffieHellman.cs
@@ -1,70 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the http://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-namespace System.Security.Cryptography
-{
-    public sealed partial class ECDiffieHellmanCng : System.Security.Cryptography.ECDiffieHellman
-    {
-        public ECDiffieHellmanCng() { }
-        public ECDiffieHellmanCng(int keySize) { }
-        public ECDiffieHellmanCng(CngKey key) { }
-#if FEATURE_ECPARAMETERS
-        public ECDiffieHellmanCng(System.Security.Cryptography.ECCurve curve) { }
-#endif
-        public System.Security.Cryptography.CngAlgorithm HashAlgorithm { get { throw null; } set { } }
-        public byte[] HmacKey { get { throw null; } set { } }
-        public System.Security.Cryptography.CngKey Key { get { throw null; } }
-        public System.Security.Cryptography.ECDiffieHellmanKeyDerivationFunction KeyDerivationFunction { get { throw null; } set { } }
-        public override int KeySize { get { throw null; } set { } }
-        public byte[] Label { get { throw null; } set { } }
-        public override System.Security.Cryptography.ECDiffieHellmanPublicKey PublicKey { get { throw null; } }
-        public byte[] SecretAppend { get { throw null; } set { } }
-        public byte[] SecretPrepend { get { throw null; } set { } }
-        public byte[] Seed { get { throw null; } set { } }
-        public bool UseSecretAgreementAsHmacKey { get { throw null; } }
-#if FEATURE_ECDH_DERIVEFROM
-        public override byte[] DeriveKeyFromHash(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, System.Security.Cryptography.HashAlgorithmName hashAlgorithm, byte[] secretPrepend, byte[] secretAppend) { throw null; }
-        public override byte[] DeriveKeyFromHmac(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, System.Security.Cryptography.HashAlgorithmName hashAlgorithm, byte[] hmacKey, byte[] secretPrepend, byte[] secretAppend) { throw null; }
-        public override byte[] DeriveKeyTls(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, byte[] prfLabel, byte[] prfSeed) { throw null; }
-#endif
-        public byte[] DeriveKeyMaterial(System.Security.Cryptography.CngKey otherPartyPublicKey) { throw null; }
-        public override byte[] DeriveKeyMaterial(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey) { throw null; }
-        public Microsoft.Win32.SafeHandles.SafeNCryptSecretHandle DeriveSecretAgreementHandle(System.Security.Cryptography.CngKey otherPartyPublicKey) { throw null; }
-        public Microsoft.Win32.SafeHandles.SafeNCryptSecretHandle DeriveSecretAgreementHandle(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey) { throw null; }
-        protected override void Dispose(bool disposing) { }
-#if FEATURE_ECPARAMETERS
-        public override System.Security.Cryptography.ECParameters ExportExplicitParameters(bool includePrivateParameters) { throw null; }
-        public override System.Security.Cryptography.ECParameters ExportParameters(bool includePrivateParameters) { throw null; }
-#endif
-        public void FromXmlString(string xml, System.Security.Cryptography.ECKeyXmlFormat format) { }
-#if FEATURE_ECPARAMETERS
-        public override void GenerateKey(System.Security.Cryptography.ECCurve curve) { }
-        public override void ImportParameters(System.Security.Cryptography.ECParameters parameters) { }
-#endif
-        public string ToXmlString(System.Security.Cryptography.ECKeyXmlFormat format) { throw null; }
-    }
-    public sealed partial class ECDiffieHellmanCngPublicKey : System.Security.Cryptography.ECDiffieHellmanPublicKey
-    {
-        private ECDiffieHellmanCngPublicKey() : base(null) { } 
-        public System.Security.Cryptography.CngKeyBlobFormat BlobFormat { get { throw null; } }
-        protected override void Dispose(bool disposing) { }
-#if FEATURE_ECPARAMETERS
-        public override System.Security.Cryptography.ECParameters ExportExplicitParameters() { throw null; }
-        public override System.Security.Cryptography.ECParameters ExportParameters() { throw null; }
-#endif
-        public static System.Security.Cryptography.ECDiffieHellmanPublicKey FromByteArray(byte[] publicKeyBlob, System.Security.Cryptography.CngKeyBlobFormat format) { throw null; }
-        public static System.Security.Cryptography.ECDiffieHellmanCngPublicKey FromXmlString(string xml) { throw null; }
-        public System.Security.Cryptography.CngKey Import() { throw null; }
-        public override string ToXmlString() { throw null; }
-    }
-    public enum ECDiffieHellmanKeyDerivationFunction
-    {
-        Hash = 0,
-        Hmac = 1,
-        Tls = 2,
-    }
-}
