+++ src/System.Security.Cryptography.Cng/ref/System.Security.Cryptography.Cng.cs
@@ -277,7 +277,9 @@ namespace System.Security.Cryptography
         public ECDiffieHellmanCng() { }
         public ECDiffieHellmanCng(int keySize) { }
         public ECDiffieHellmanCng(System.Security.Cryptography.CngKey key) { }
+#if FEATURE_ECPARAMETERS
         public ECDiffieHellmanCng(System.Security.Cryptography.ECCurve curve) { }
+#endif
         public System.Security.Cryptography.CngAlgorithm HashAlgorithm { get { throw null; } set { } }
         public byte[] HmacKey { get { throw null; } set { } }
         public System.Security.Cryptography.CngKey Key { get { throw null; } }
@@ -289,19 +291,27 @@ namespace System.Security.Cryptography
         public byte[] SecretPrepend { get { throw null; } set { } }
         public byte[] Seed { get { throw null; } set { } }
         public bool UseSecretAgreementAsHmacKey { get { throw null; } }
+#if FEATURE_ECDH_DERIVEFROM
         public override byte[] DeriveKeyFromHash(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, System.Security.Cryptography.HashAlgorithmName hashAlgorithm, byte[] secretPrepend, byte[] secretAppend) { throw null; }
         public override byte[] DeriveKeyFromHmac(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, System.Security.Cryptography.HashAlgorithmName hashAlgorithm, byte[] hmacKey, byte[] secretPrepend, byte[] secretAppend) { throw null; }
+#endif
         public byte[] DeriveKeyMaterial(System.Security.Cryptography.CngKey otherPartyPublicKey) { throw null; }
         public override byte[] DeriveKeyMaterial(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey) { throw null; }
+#if FEATURE_ECDH_DERIVEFROM
         public override byte[] DeriveKeyTls(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey, byte[] prfLabel, byte[] prfSeed) { throw null; }
+#endif
         public Microsoft.Win32.SafeHandles.SafeNCryptSecretHandle DeriveSecretAgreementHandle(System.Security.Cryptography.CngKey otherPartyPublicKey) { throw null; }
         public Microsoft.Win32.SafeHandles.SafeNCryptSecretHandle DeriveSecretAgreementHandle(System.Security.Cryptography.ECDiffieHellmanPublicKey otherPartyPublicKey) { throw null; }
         protected override void Dispose(bool disposing) { }
+#if FEATURE_ECPARAMETERS
         public override System.Security.Cryptography.ECParameters ExportExplicitParameters(bool includePrivateParameters) { throw null; }
         public override System.Security.Cryptography.ECParameters ExportParameters(bool includePrivateParameters) { throw null; }
+#endif
         public void FromXmlString(string xml, System.Security.Cryptography.ECKeyXmlFormat format) { }
+#if FEATURE_ECPARAMETERS
         public override void GenerateKey(System.Security.Cryptography.ECCurve curve) { }
         public override void ImportParameters(System.Security.Cryptography.ECParameters parameters) { }
+#endif
         public string ToXmlString(System.Security.Cryptography.ECKeyXmlFormat format) { throw null; }
     }
     public sealed partial class ECDiffieHellmanCngPublicKey : System.Security.Cryptography.ECDiffieHellmanPublicKey
@@ -309,8 +319,10 @@ namespace System.Security.Cryptography
         internal ECDiffieHellmanCngPublicKey() { }
         public System.Security.Cryptography.CngKeyBlobFormat BlobFormat { get { throw null; } }
         protected override void Dispose(bool disposing) { }
+#if FEATURE_ECPARAMETERS
         public override System.Security.Cryptography.ECParameters ExportExplicitParameters() { throw null; }
         public override System.Security.Cryptography.ECParameters ExportParameters() { throw null; }
+#endif
         public static System.Security.Cryptography.ECDiffieHellmanPublicKey FromByteArray(byte[] publicKeyBlob, System.Security.Cryptography.CngKeyBlobFormat format) { throw null; }
         public static System.Security.Cryptography.ECDiffieHellmanCngPublicKey FromXmlString(string xml) { throw null; }
         public System.Security.Cryptography.CngKey Import() { throw null; }
