+++ src/System.Security.Cryptography.Cng/ref/System.Security.Cryptography.Cng.csproj
@@ -11,7 +11,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Security.Cryptography.Cng.cs" />
-    <Compile Include="System.Security.Cryptography.Cng.ECDiffieHellman.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(IsPartialFacadeAssembly)' == 'true'">
     <Reference Include="mscorlib" />
