+++ src/System.ServiceModel.Syndication/ref/System.ServiceModel.Syndication.cs
@@ -360,7 +360,6 @@ namespace System.ServiceModel.Syndication
         public System.Collections.ObjectModel.Collection<System.ServiceModel.Syndication.SyndicationPerson> Contributors { get { throw null; } }
         public System.ServiceModel.Syndication.TextSyndicationContent Copyright { get { throw null; } set { } }
         public System.ServiceModel.Syndication.TextSyndicationContent Description { get { throw null; } set { } }
-        public System.ServiceModel.Syndication.SyndicationLink Documentation { get { throw null; } set { } }
         public System.ServiceModel.Syndication.SyndicationElementExtensionCollection ElementExtensions { get { throw null; } }
         public string Generator { get { throw null; } set { } }
         public string Id { get { throw null; } set { } }
@@ -369,10 +368,6 @@ namespace System.ServiceModel.Syndication
         public string Language { get { throw null; } set { } }
         public System.DateTimeOffset LastUpdatedTime { get { throw null; } set { } }
         public System.Collections.ObjectModel.Collection<System.ServiceModel.Syndication.SyndicationLink> Links { get { throw null; } }
-        public System.Collections.ObjectModel.Collection<string> SkipDays { get { throw null; } }
-        public System.Collections.ObjectModel.Collection<int> SkipHours { get { throw null; } }
-        public System.ServiceModel.Syndication.SyndicationTextInput TextInput { get { throw null; } set { } }
-        public System.TimeSpan? TimeToLive { get { throw null; } set { } }
         public System.ServiceModel.Syndication.TextSyndicationContent Title { get { throw null; } set { } }
         public virtual System.ServiceModel.Syndication.SyndicationFeed Clone(bool cloneItems) { throw null; }
         protected internal virtual System.ServiceModel.Syndication.SyndicationCategory CreateCategory() { throw null; }
@@ -396,9 +391,7 @@ namespace System.ServiceModel.Syndication
     {
         protected SyndicationFeedFormatter() { }
         protected SyndicationFeedFormatter(System.ServiceModel.Syndication.SyndicationFeed feedToWrite) { }
-        public System.ServiceModel.Syndication.TryParseDateTimeCallback DateTimeParser { get { throw null; } set { } }
         public System.ServiceModel.Syndication.SyndicationFeed Feed { get { throw null; } }
-        public System.ServiceModel.Syndication.TryParseUriCallback UriParser { get { throw null; } set { } }
         public abstract string Version { get; }
         public abstract bool CanRead(System.Xml.XmlReader reader);
         protected internal static System.ServiceModel.Syndication.SyndicationCategory CreateCategory(System.ServiceModel.Syndication.SyndicationFeed feed) { throw null; }
@@ -561,14 +554,6 @@ namespace System.ServiceModel.Syndication
         protected internal virtual void WriteAttributeExtensions(System.Xml.XmlWriter writer, string version) { }
         protected internal virtual void WriteElementExtensions(System.Xml.XmlWriter writer, string version) { }
     }
-    public partial class SyndicationTextInput
-    {
-        public SyndicationTextInput() { }
-        public string Description { get { throw null; } set { } }
-        public System.ServiceModel.Syndication.SyndicationLink Link { get { throw null; } set { } }
-        public string Name { get { throw null; } set { } }
-        public string Title { get { throw null; } set { } }
-    }
     public static partial class SyndicationVersions
     {
         public const string Atom10 = "Atom10";
@@ -590,8 +575,6 @@ namespace System.ServiceModel.Syndication
         Plaintext = 0,
         XHtml = 2,
     }
-    public delegate bool TryParseDateTimeCallback(System.ServiceModel.Syndication.XmlDateTimeData data, out System.DateTimeOffset dateTimeOffset);
-    public delegate bool TryParseUriCallback(System.ServiceModel.Syndication.XmlUriData data, out System.Uri uri);
     public partial class UrlSyndicationContent : System.ServiceModel.Syndication.SyndicationContent
     {
         protected UrlSyndicationContent(System.ServiceModel.Syndication.UrlSyndicationContent source) { }
@@ -617,13 +600,6 @@ namespace System.ServiceModel.Syndication
         protected internal virtual void WriteAttributeExtensions(System.Xml.XmlWriter writer, string version) { }
         protected internal virtual void WriteElementExtensions(System.Xml.XmlWriter writer, string version) { }
     }
-    public partial struct XmlDateTimeData
-    {
-        private object _dummy;
-        public XmlDateTimeData(string dateTimeString, System.Xml.XmlQualifiedName elementQualifiedName) { throw null; }
-        public string DateTimeString { get { throw null; } }
-        public System.Xml.XmlQualifiedName ElementQualifiedName { get { throw null; } }
-    }
     public partial class XmlSyndicationContent : System.ServiceModel.Syndication.SyndicationContent
     {
         protected XmlSyndicationContent(System.ServiceModel.Syndication.XmlSyndicationContent source) { }
@@ -639,14 +615,5 @@ namespace System.ServiceModel.Syndication
         public TContent ReadContent<TContent>(System.Runtime.Serialization.XmlObjectSerializer dataContractSerializer) { throw null; }
         public TContent ReadContent<TContent>(System.Xml.Serialization.XmlSerializer serializer) { throw null; }
         protected override void WriteContentsTo(System.Xml.XmlWriter writer) { }
-    }
-    public partial struct XmlUriData
-    {
-        private object _dummy;
-        private int _dummyPrimitive;
-        public XmlUriData(string uriString, System.UriKind uriKind, System.Xml.XmlQualifiedName elementQualifiedName) { throw null; }
-        public System.Xml.XmlQualifiedName ElementQualifiedName { get { throw null; } }
-        public System.UriKind UriKind { get { throw null; } }
-        public string UriString { get { throw null; } }
     }
 }
