+++ src/System.Threading.AccessControl/ref/System.Threading.AccessControl.csproj
@@ -6,7 +6,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Threading.AccessControl.cs" />
-    <Compile Include="System.Threading.AccessControl.Manual.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetsNetFx)' == 'true'">
     <Reference Include="mscorlib" />
